% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust_playback.r
\name{adjust_playback_time}
\alias{adjust_playback_time}
\title{Modify playback time of video}
\usage{
adjust_playback_time(scale_factor = 1, input, output_dir = ".",
  output = "test.mp4", overwrite = FALSE, ffmpeg = NA)
}
\arguments{
\item{scale_factor}{multiplicative factor changes duration of
video playback. See details.}

\item{input}{character, path to video file}

\item{output_dir}{character, output directory, default is working
directory}

\item{output}{character, output file name}

\item{overwrite}{logical, default is \code{TRUE}}

\item{ffmpeg}{A character string with path to install directory
for ffmpeg.  This argument is only needed if ffmpeg has not been
added to your path variable on your computer.  For Windows
machines, path must point to ffmpeg.exe, typically located in
bin folder.  For example,}
}
\value{
One video animation will be written to \code{output_dir}
}
\description{
Speed up or slow down playback of video using FFmpeg software.
}
\details{
\code{adjust_playback_time} is a helper function that adjusts
  duration of video.  \code{scale_factor} controls the magnitude
  of speed-up or slow-down by modifying the presentation timestamp
  of each video frame. For more information, see
  Values of \code{scale_factor} < 1 speed up playback and >1 slow
  down video.
}
\author{
Todd Hayden
}
