% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animateFrames.r
\name{animateFrames}
\alias{animateFrames}
\title{Use FFMPEG to create an animated video of interpolated fish movements}
\usage{
animateFrames(procObj, recs, outDir, background = NULL,
  backgroundYlim = c(41.48, 45.9), backgroundXlim = c(-84, -79.5),
  ffmpeg = NA, recColNames = list(latitudeCol = "deploy_lat", longitudeCol =
  "deploy_long", deploy_timestampCol = "deploy_date_time", recover_timestampCol
  = "recover_date_time"))
}
\arguments{
\item{procObj}{A data frame created by \code{\link{animate}} function}

\item{recs}{A data frame containing at least four columns with 
receiver 'lat', 'lon', 'deploy_timestamp', and 
'recover_timestamp'. Default column names match GLATOS standard receiver 
location file \cr(e.g., 'GLATOS_receiverLocations_yyyymmdd.csv'), but 
column names can also be specified with \code{recColNames}.}

\item{outDir}{A character string with file path to directory where 
individual frames for animations will be written.}

\item{background}{object of class \code{SpatialPolygonsDataFrame} 
to be used as background of each frame.}

\item{backgroundYlim}{vector of two values specifying the min/max values 
for y-scale of plot. Units are same as background argument.}

\item{backgroundXlim}{vector of two values specifying the min/max values 
for x-scale of plot. Units are same as background argument.}

\item{ffmpeg}{A character string with path to install directory for ffmpeg. 
This argument is only needed if ffmpeg has not been added to your 
path variable on your computer.  For Windows machines, path must point 
to ffmpeg.exe.  For example 'c:\\path\\to\\ffmpeg\\bin\\ffmpeg.exe'}

\item{recColNames}{A list with names of required columns in 
  \code{recs}: 
\itemize{
 \item \code{latitudeCol} is a character string with the name of the column
    containing latitude of the receiver (typically 'deploy_lat' for 
    GLATOS standard detection export data). 
 \item \code{longitudeCol} is a character string with the name of the column
    containing longitude of the receiver (typically 'deploy_long' for 
    GLATOS standard detection export data).
 \item \code{deploy_timestampCol} is a character string with the name of 
    the column containing datetime stamps for receier deployments (MUST be 
    of class 'POSIXct'; typically 'deploy_date_time'for GLATOS standard 
    detection export data). 
 \item \code{recover_timestampCol} is a character string with the name of 
    the column containing datetime stamps for receier recover (MUST be of 
    class 'POSIXct'; typically 'recover_date_time'for GLATOS standard 
    detection export data). 
}}
}
\value{
video animation and individual frames, numbered sequentially
}
\description{
Plot actual and interpolated time-series positions (e.g., detections), 
export frames as png files, and use FFMPEG to stitch each 
frame together into a video (mp4) animation.
}
\examples{
library(glatos)

#example detection data
data(walleye_detections) 
head(walleye_detections)

#example receiver location data
data(recLoc_example) 
head(recLoc_example)

#call with defaults; linear interpolation
pos1 <- animate(walleye_detections)

#make sure ffmpeg is installed before calling animateFrames
# and add path to 'ffmpeg' on next line (or add path to PATH VAR)
animateFrames(pos1, recs=recLoc_example, ffmpeg=NA)

}
\author{
Todd Hayden
}
