% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animatePath.r
\name{animatePath}
\alias{animatePath}
\title{Create an animated video of spatiotemporal path data}
\usage{
animatePath(proc_obj, recs, plot_control = NULL, out_dir = getwd(),
  background = NULL, background_ylim = c(41.48, 45.9),
  background_xlim = c(-84, -79.5), ffmpeg = NA,
  ani_name = "animation.mp4", frame_delete = TRUE, animate = TRUE)
}
\arguments{
\item{proc_obj}{A data frame created by \code{\link{interpolatePath}} 
function.}

\item{recs}{A data frame containing at least four columns with
receiver 'deploy_lat', 'deploy_long', 'deploy_date_time', and
'recover_date_time'. Other columns in object will be ignored.
Default column names match GLATOS standard receiver location file
\cr(e.g., 'GLATOS_receiverLocations_yyyymmdd.csv').}

\item{plot_control}{An optional data frame with four columns ('animal_id', 'type', 
  'color', and 'marker', 'marker_cex') that specify the plot symbols and colors for 
  each animal and position type. See examples below for an example.
\itemize{
  \item \code{animal_id} contains the unique identifier of individual animals and 
       corresponds to 'animal_id' column in 'dtc'. 
  \item \code{type} indicates if the options should be applied to observed
    positions (detections; 'real') or interpolated positions 
    ('inter').
  \item \code{color} contains the marker color to be plotted for each 
    animal and position type.  
  \item \code{marker} contains the marker style to be plotted for each
    animal and position type. Passed to \code{par()$pch}.
  \item \code{marker_cex} contains the marker size to be plotted for each
    animal and position type. Passed to \code{par()$cex}.
}}

\item{out_dir}{A character string with file path to directory where 
individual frames for animations will be written. Default is working directory.}

\item{background}{An optional object of class \code{SpatialPolygonsDataFrame} 
to be used as background of each frame. Default is a simple polygon
of the Great Lakes (\code{greatLakesPoly}) included in the 'glatos' 
package.}

\item{background_ylim}{vector of two values specifying the min/max values 
for y-scale of plot. Units are same as background argument.}

\item{background_xlim}{vector of two values specifying the min/max values 
for x-scale of plot. Units are same as background argument.}

\item{ffmpeg}{A character string with path to install directory for ffmpeg. 
This argument is only needed if ffmpeg has not been added to your 
path variable on your computer.  For Windows machines, path must point 
to ffmpeg.exe.  For example, 'c:\\path\\to\\ffmpeg\\bin\\ffmpeg.exe'}

\item{ani_name}{Name of animation (character string)}

\item{frame_delete}{Boolean.  Default (TRUE) delete individual
image frames after animation is created}

\item{animate}{Boolean. Default (TRUE) creates video animation}
}
\value{
Sequentially-numbered png files (one for each frame) and 
  one mp4 file will be written to \code{out_dir}.
}
\description{
Create a set of frames (png image files) showing geographic location data
(e.g., detections of tagged fish or interpolated path data) at discrete 
points in time and stitch frames into a video animation (mp4 file).
}
\examples{
library(glatos)
#example detection data
data(walleye_detections) 
head(walleye_detections)

#example receiver location data
data(recLoc_example) 
head(recLoc_example)

#call with defaults; linear interpolation
pos1 <- interpolatePath(walleye_detections)

#make sure ffmpeg is installed before calling animatePath
# and if you have not added path to 'ffmpeg.exe' to your Windows PATH 
# environment variable then you'll need to do that  
# or set path to 'ffmpeg.exe' using the 'ffmpeg' input argument
myDir <- paste0(getwd(),"/frames")
animatePath(pos1, recs=recLoc_example, outDir=myDir)

 
}
\author{
Todd Hayden
}
