% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summ-breakdown.r
\name{breakdown}
\alias{breakdown}
\title{Breakdown and summarize GLATOS detections by individual and location.}
\usage{
breakdown(dtc, by_col = list(individual = c("animal_id"), location =
  c("glatos_array", "station_no")))
}
\arguments{
\item{dtc}{A GLATOS detection data frame (class \code{glatos_detections}).}

\item{by_col}{A list with two elements. The first element, named 'individual'
is a vector of character strings containing names of columns in dtc that
together identify each unique animal or group of animals. The second
element, named 'location', is a vector of character strings containing
names of columns in dtc that together identify each unique location or
group of locations.}
}
\value{
A list of two data frames. The first named 'individuals' and the 
 second named 'locations'. Each data frame one row per unique combination
 of 'location' or 'individual' variables and then the count of records, 
 count of unique oppposing groups, and min and max detection timestamps.
}
\description{
Breakdown and summarize GLATOS detections by individual and location.
}
\details{
TBD
}
\examples{
#get path to example detection file
det_file <- system.file("extdata", 
  "walleye_detections.zip",package="glatos")
det_file <- unzip(det_file, "walleye_detections.csv")
det <- read_glatos_detections(det_file)
bd <- breakdown(det)
bd <- breakdown(det, by_col = list(individual = c("release_location"),
                location = c("glatos_array", "station_no")))
bd <- breakdown(det, by_col = list(individual = NULL,
                location = c("glatos_array", "station_no")))
                
}
\author{
C. Holbrook (cholbrook@usgs.gov)
}
