% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim-crw_in_polygon.r
\name{check_cross_boundary}
\alias{check_cross_boundary}
\title{Check if track crosses polygon boundary}
\usage{
check_cross_boundary(path, boundary, EPSG)
}
\description{
Check if track crosses polygon boundary
}
