% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-check_dependencies.r
\name{check_dependencies}
\alias{check_dependencies}
\title{Check system dependencies necessary for package 'glatos'}
\usage{
check_dependencies()
}
\value{
A list:
}
\description{
Checks to make sure external dependencies required by functions in
'glatos' are installed and available to R.
}
\details{
Function returns a list of tests for all external
  dependencies needed for all functions in 'glatos'.
}
\author{
Todd Hayden, Tom Binder, Chris Holbrook
}
