% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-vdat.r
\name{check_vdat}
\alias{check_vdat}
\title{Check path to Innovasea program \code{vdat.exe}}
\usage{
check_vdat(vdat_exe_path = NULL)
}
\arguments{
\item{vdat_exe_path}{The full path to \code{vdat.exe}. If \code{NULL}
(default) then the path to \code{vdat.exe} must be in the PATH environment
variable of the system.}
}
\value{
Character string with command for calling \code{vdat.exe} via
\code{system2}'s \code{command} argument.
}
\description{
Check path to Innovasea program \code{vdat.exe}
}
\examples{
\dontrun{

# use Windows system PATH variable
check_vdat()


# use path to directory containing vdat.exe
check_vdat(vdat_exe_path = "C:/Program Files/Innovasea/Fathom Connect")


# use full path to vdat.exe
check_vdat(vdat_exe_path = "C:/Program Files/Innovasea/Fathom Connect/vdat.exe")
}

}
