% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-vue.r
\name{check_vue}
\alias{check_vue}
\title{Check path to Innovasea program VUE.exe}
\usage{
check_vue(vue_exe_path = NULL)
}
\arguments{
\item{vue_exe_path}{The full path to \code{VUE.exe}. If \code{NULL}
(default) then the path to VUE.exe must be in the PATH environment variable
of the system.}
}
\value{
Character string with command for calling VUE.exe via
\code{system2}'s \code{command} argument.
}
\description{
Check path to Innovasea program VUE.exe
}
\examples{
\dontrun{

# use Windows system PATH variable
check_vue()


# use path to directory containing VUE.exe
check_vue(vue_exe_path = "C:/Program Files (x86)/VEMCO/VUE")


# use full path to VUE.exe
check_vue(vue_exe_path = "C:/Program Files (x86)/VEMCO/VUE/VUE.exe")
}

}
