% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condense_events.R
\name{condense_events}
\alias{condense_events}
\title{Classify detection data into events}
\usage{
condense_events(dtc, loc_col = "glatos_array", timeSep = Inf)
}
\arguments{
\item{dtc}{An object of class \code{glatos_detections} of data
frame containing spatiotemporal data with a location column
(typically 'glatos_array' or 'station'), 'animal_id' and
'detection_timestamp_utc' columns.}

\item{loc_col}{Name of column in 'dtc' that specifies location column
'}

\item{timeSep}{Amount of time (in seconds) that must pass between 
sequential detections on the same receiver (or group of receivers, 
depending on specified location) before that detection is considered to 
belong to a new detection event. The default value \code{Inf}, will not 
define events based on elapsed time (only when location changes).}
}
\value{
Input dataframe with columns containing flags:
 \item{arrival}{Flag (0 or 1) representing the first detection in each event.} 
 \item{depart}{Flag (0 or 1) representing the last detection in each event.} 
 \item{event}{integer representing the event number}
 \item{TimeDiff}{Lagged time difference in seconds between successive detections.}
 \item{newfsh}{Flag representing first detection of each fish}
}
\description{
Reduce detection data from an acoustic telemetry receiver into discrete 
  detection events, defined by movement between receivers (or receiver 
  groups, depending on location), or sequential detections at the same 
  location that are separated by a user-defined threshold period of time.
}
\examples{

# load detection data
det_file <- system.file("extdata", "walleye_detections.zip", package = "glatos")
det_file <- unzip(det_file, "walleye_detections.csv")
dtc <- read_glatos_detections(det_file)

dtx <- condense_events(dtc, timeSep = 3600)
}
\author{
T. R. Binder, Todd Hayden
}
