% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-convert_glatos_to_att.r
\name{convert_glatos_to_att}
\alias{convert_glatos_to_att}
\title{Convert detections and receiver metadata to a format that ATT accepts.}
\usage{
convert_glatos_to_att(detectionObj, receiverObj, crs = sf::st_crs(4326))
}
\arguments{
\item{detectionObj}{A \code{glatos_detections} object (e.g., created by
\link{read_glatos_detections}) or a \code{data.frame} containing required columns
(see \link{glatos_detections}).}

\item{receiverObj}{A \code{glatos_receivers} object (e.g., created by
\link{read_glatos_receivers}) or a \code{data.frame} containing required columns
(see \link{glatos_receivers}).}

\item{crs}{an object of class \code{crs} (see \link[=st_crs]{sf::st_crs}) with
geographic coordinate system for all spatial information
(latitude/longitude). If none provided or \code{crs} is not recognized,
defaults to WGS84.}
}
\value{
a list of 3 tibbles containing tag detections, tag metadata,
and station metadata, to be ingested by VTrack/ATT
}
\description{
Convert \code{glatos_detections} and \code{glatos_receiver} objects to \code{ATT} for
compatibility with the Animal Tracking Toolbox
\url{https://github.com/vinayudyawer/ATT}, now part of \code{VTrack}
\url{https://github.com/RossDwyer/VTrack}.
}
\details{
This function takes 2 lists containing detection and reciever data
and transforms them into one list containing 3 \code{tibble} objects. The input
that AAT uses to get this data product is located here:
\url{https://github.com/vinayudyawer/ATT/blob/master/README.md} and our mappings
are found here:
\url{https://github.com/ocean-tracking-network/glatos/issues/75#issuecomment-982822886}
in a comment by Ryan Gosse.

Note that the \code{Tag.Detections} element of the output can contain
fewer records than \code{detectionObj} because detections are omitted if they
do not match a station deployment-recovery interval in \code{receiverObj}.
For example, in the walleye data example, \code{walleye_detections} contains
7180 rows but \code{Tag.Detectons} output only contains 7083 rows.
}
\examples{

#--------------------------------------------------
# EXAMPLE #1 - loading from the vignette data

library(glatos)
wal_det_file <- system.file("extdata", "walleye_detections.csv",
  package = "glatos"
)
walleye_detections <- read_glatos_detections(wal_det_file) # load walleye data

rec_file <- system.file("extdata", "sample_receivers.csv",
  package = "glatos"
)
rcv <- read_glatos_receivers(rec_file) # load receiver data

ATTdata <- convert_glatos_to_att(walleye_detections, rcv)
}
\author{
Ryan Gosse
}
