% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-convert_otn_erddap_to_att.r
\name{convert_otn_erddap_to_att}
\alias{convert_otn_erddap_to_att}
\title{Convert detections and receiver metadata to a format that 
ATT (https://github.com/vinayudyawer/ATT) accepts.}
\usage{
convert_otn_erddap_to_att(detectionObj, erdTags, erdRcv, erdAni)
}
\arguments{
\item{detectionObj}{a list from \code{read_glatos_detections}}

\item{erdTags}{a list from the OTN ERDDAP of tags}

\item{erdRcv}{a list from the OTN ERDDAP of receivers}

\item{erdAni}{a list from the OTN ERDDAP of animals}
}
\value{
a list of 3 tibble::tibbles containing tag dectections, tag metadata, and
station metadata, to be injested by VTrack/ATT
}
\description{
Convert detections and receiver metadata to a format that 
ATT (https://github.com/vinayudyawer/ATT) accepts.
}
\details{
This function takes 4 lists containing detection, and
ERDDAP data from the tags receivers and animals tables, and transforms them into 3 \code{tibble::tibble} objects 
inside of a list. The input that AAT uses to get this data product
is located here: https://github.com/vinayudyawer/ATT/blob/master/README.md
and our mappings are found here: https://gitlab.oceantrack.org/GreatLakes/glatos/issues/83
in a comment by Ryan Gosse. The OTN ERDDAP instance is here: https://members.oceantrack.org/erddap/tabledap/index.html
but please note that this only contains public data.
}
\examples{

#--------------------------------------------------
# EXAMPLE #1 - loading from the OTN ERDDAP + vignettes

library(glatos)
animals <- read.csv("/path/to/animal/csv/otn_aat_animals.csv") # load the CSVs from ERDDAP
tags <- read.csv("/path/to/tag/csv/otn_aat_tags.csv")
stations <- read.csv("/path/to/station/csv/otn_aat_stations.csv")

animals <- animals[-1,] # First row is garbage (blank or metadata about the column) so we can cut it
tags <- tags[-1,]
stations <- stations[-1,]

shrk_det_file <- system.file("extdata", "blue_shark_detections.csv",
     package = "glatos")
blue_shark_detections <- read_glatos_detections(shrk_det_file) # load shark data

ATTdata <- convert_otn_erddap_to_att(blue_shark_detections, tags, stations, animals)
}
\author{
Ryan Gosse
}
