% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim-crw_in_polygon.r
\name{crosses_boundary}
\alias{crosses_boundary}
\title{Check if path segments cross polygon boundary}
\usage{
crosses_boundary(path, boundary)
}
\arguments{
\item{path}{A two-column matrix of points, representing x and y,
respectively, along a path.}

\item{boundary}{An \code{sf} object containing \code{MULTILINESTRING} geometry
representing a polygon boundary.}
}
\value{
A logical vector with an element for each 'step' in \code{path} that
indicates if that step crosses \code{boundary} (TRUE) or not (FALSE).
}
\description{
Internal function used in \code{\link[=crw_in_polygon]{crw_in_polygon()}} to determine if
(and identify which) line segments cross polygon boundaries (e.g., steps
onto land or over a peninsula).
}
\details{
Check if each segment in a path crosses polygon boundary
}
\examples{

# Example 1

# make path
path <- matrix(c(0:6, rep(3, 7)), ncol = 2)

# make polygon
poly <- matrix(c(0,0, 6,0, 3,6, 0,0), ncol = 2, byrow = TRUE)

plot(poly, type = "l")
lines(path, type = "o", col = "red")

poly <- sf::st_linestring(poly)

crosses_boundary(path, poly)

#Example 2

# make path
path <- matrix(c(0,1,5,6, rep(3, 4)), ncol = 2)

# make polygon
poly <- matrix(c(0,0, 6,0, 3,6, 0,0), ncol = 2, byrow = TRUE)

plot(poly, type = "l")
lines(path, type = "o", col = "red")

poly <- sf::st_linestring(poly)

crosses_boundary(path, poly)

}
