% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-impute_detections.r
\name{estimate_delay}
\alias{estimate_delay}
\title{estimate fixed delay}
\usage{
estimate_delay(dtc, tags = NA, recs = NA, tolerance = 1)
}
\description{
function to estimate time between detections
}
