% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-impute_detections.r
\name{estimate_fixed_delay_density}
\alias{estimate_fixed_delay_density}
\title{estimate fixed delay from detection data using density method}
\usage{
estimate_fixed_delay_density(dtc, ...)
}
\arguments{
\item{dtc}{A data frame with detection data}

\item{...}{Optional named list of arguments passed to \code{density}.}
}
\description{
estimate fixed delay from detection data using density method
}
\details{
This only accepts data for a single tag-receiver pair.
}
