% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-impute_detections.r
\name{estimate_fixed_delay_lm}
\alias{estimate_fixed_delay_lm}
\title{estimate fixed delay from detection data using regression method}
\usage{
estimate_fixed_delay_lm(dtc, exp_delay, ...)
}
\arguments{
\item{dtc}{A data frame with detection data}

\item{...}{Optional named list of arguments passed to \code{lm}.}
}
\description{
estimate fixed delay from detection data using regression method
}
\details{
This only accepts data for a single tag-receiver pair.
}
