% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filt-false_detections.r
\name{false_detections}
\alias{false_detections}
\title{False detection filter}
\usage{
false_detections(det, tf, min_lag_col = "min_lag", show_plot = FALSE, ...)
}
\arguments{
\item{det}{A \code{glatos_detections} object (e.g., produced by
\link{read_glatos_detections}).

\emph{OR:} A data frame with one column containing 'min_lag' which for each
detection record, is the smallest time (in seconds) to the next closest
detection (either previous or subsequent) of the same transmitter on the
same receiver. The name of the column containing 'min_lag' can be specified
via \code{min_lag_col}; see below).

\emph{OR} (\emph{if \code{min_lag} is missing}) A data farme containing
detection data with the four columns described below. In that case,
\code{min_lag} will be calculated using \link{min_lag}).
\describe{
\item{detection_timestamp_utc}{Detection timestamps; MUST be of class
POSIXct.}
\item{transmitter_codespace}{A character string with transmitter code space
(e.g., "A69-1061" for Vemco PPM coding").}
\item{transmitter_id}{A character string with transmitter ID code (e.g.,
"1363" for Vemco PPM coding").}
\item{receiver_sn}{A character vector with unique receiver serial number.}
}}

\item{tf}{A number indicating the time threshold (in seconds; e.g.,
Pincock's (2012) "short interval") for identifying possible false
detections.}

\item{min_lag_col}{A character string containing the name of the column
in \code{det} that contains 'min_lag'.}

\item{show_plot}{Indicates if a plot should be displayed showing the
proportion of detections that exceed min_lag from min_lag = 1 to
min_lag = 5 * tf.}

\item{...}{Additional arguments passed to \code{\link[graphics:plot]{graphics::plot()}}.}
}
\value{
A data frame consisting of \code{det} with an additional
column 'passed_filter' indicating if each detection did (1) or did not (0)
pass the criteria.
}
\description{
Identify possible false detections based on "short interval" criteria (e.g.,
GLATOS 'min_lag') .
}
\details{
Detections are identified as potentially false when
\code{min_lag > tf}.

A new column (\code{passed_filter}), indicating if each record (row)
passed the filter, is added to the input data frame.

This function was written specifically with GLATOS standard
detection export in mind, but if \code{min_lag} is absent and
\code{min_lag_col} is not specified, then \code{min_lag} will be calculated
using \link{min_lag}.

A common rule of thumb for choosing \code{tf} for VEMCO PPM encoded
transmitters is 30 times the nominal delay (e.g., 3600 s for a transmitter
with a 120 s nominal delay) - see Pincock (2012).

When \code{show_plot = TRUE} then the plot may be used to assess
sensitivity of the proportion of detections removed to the choice of
\code{tf}.
}
\examples{
# get path to example detection file
det_file <- system.file("extdata", "walleye_detections.csv",
  package = "glatos"
)
det <- read_glatos_detections(det_file)

det <- false_detections(det, 3600)
head(det)

# plot sensitivity to tf
det <- false_detections(det, 3600, show_plot = TRUE)

}
\references{
Pincock, D.G., 2012. False detections: what they are and how to remove them
from detection data. Vemco Division, Amirix Systems Inc., Halifax,
Nova Scotia.
\cr \url{http://www.vemco.com/pdf/false_detections.pdf}

Simpfendorfer, C.A., Huveneers, C., Steckenreuter, A., Tattersall, K.,
Hoenner, X., Harcourt, R. and Heupel, M.R., 2015. Ghosts in the data:
false detections in VEMCO pulse position modulation acoustic telemetry
monitoring equipment. Animal Biotelemetry, 3(1), p.55.
\cr \url{https://animalbiotelemetry.biomedcentral.com/articles/10.1186/s40317-015-0094-z}
}
\seealso{
\code{\link[=min_lag]{min_lag()}}
}
\author{
T. R. Binder, edited by A. Dini
}
