% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filt-falseDetectionFilter.r
\name{falseDetectionFilter}
\alias{falseDetectionFilter}
\title{False detection filter}
\usage{
falseDetectionFilter(detections, type = "GLATOS", tf = 3600,
  minLagCol = "min_lag", detColNames = list())
}
\arguments{
\item{detections}{A data frame containing detection data (e.g., from the 
standard GLATOS detection export file '*_detectionsWithLocs.csv'). 
Must contain the column \code{min_lag} or equivalent (specified by 
\code{minLagCol}). \code{min_lag} is loosely based on the the 
"short interval" method described by Pincock (2012). In this case 
(GLATOS), it is defined for each detection as the shortest interval 
(in seconds) between either the previous or next detection (whichever 
is closest) of the same transmitter on the same receiver.}

\item{type}{A character string that contains the type of data that is being 
passed in, for example, "OTN", "GLATOS", or "sample". Default is "GLATOS".}

\item{tf}{A numeric scalar indicating the time threshold (in seconds; e.g., 
Pincock's (2012) "short interval") for identifying possible false 
detections.}

\item{minLagCol}{A character string containing the name of the column 
in \code{detections} that contains 'min_lag'.}

\item{detColNames}{An optional list that contains the user-defined column
names}
}
\value{
A data frame consisting of \code{detections} with an additional 
  column 'passedFilter' indicating if each detection did (1) or did not (0) 
  pass the criteria.
}
\description{
Identify possible false detections based on "short interval" criteria
}
\details{
The min_lag column should be in \code{detections} but if not, 
 program will add in using getMinLag() function.

Detections are identified as potentially false when 
  \code{min_lag > tf}.

A new column (\code{passedFilter}), indicating if each record (row) passed 
the filter, is added to the input data frame.

A common rule of thumb for choosing tf for VEMCO PPM encoded transmitters 
  is 30 times the nominal delay (e.g., 3600 s for a transmitter with a 
  120 s nominal delay) - see Pincock (2012).
}
\examples{
library(glatos)

# example - GLATOS data
data("walleye_detections") #example data

head(walleye_detections)

dtx <- falseDetectionFilter(walleye_detections, 3600)
head(dtx)
  
}
\references{
Pincock, D.G., 2012. False detections: what they are and how to remove them 
    from detection data. Vemco Division, Amirix Systems Inc., Halifax, 
    Nova Scotia.
    \cr \url{http://www.vemco.com/pdf/false_detections.pdf}

Simpfendorfer, C.A., Huveneers, C., Steckenreuter, A., Tattersall, K., 
    Hoenner, X., Harcourt, R. and Heupel, M.R., 2015. Ghosts in the data: 
    false detections in VEMCO pulse position modulation acoustic telemetry 
    monitoring equipment. Animal Biotelemetry, 3(1), p.55.
    \cr \url{https://animalbiotelemetry.biomedcentral.com/articles/10.1186/s40317-015-0094-z}
}
\author{
T. R. Binder, edited by A. Dini
}
