% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{flynn_island_polygon}
\alias{flynn_island_polygon}
\title{An sf POLYGON object with coastline of Flynn Island}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 1 rows and 2 columns.
}
\usage{
flynn_island_polygon
}
\description{
An sf POLYGON object with coastline of Flynn Island; and 
island within Higgins Lake, Michigan. 
Used as an example of a polygon representing a body of land (as opposed to 
water body).
}
\author{
Chris Holbrook
}
\keyword{datasets}
