% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summ-residence_index.r
\name{get_days}
\alias{get_days}
\title{Determines which calculation method to use for the residency index.}
\usage{
get_days(dets, calculation_method = "kessel",
  time_interval_size = "1 day")
}
\arguments{
\item{dets}{- data frame pulled from the detection events}

\item{calculation_method}{- determines which method above will be used to
count total time and location time}

\item{time_interval_size}{- size of time interval}
}
\description{
Wrapper method for the calulation methods above.
}
