% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-vdat.r
\name{get_local_vdat_schema}
\alias{get_local_vdat_schema}
\title{Get schema from local installation of Innovsea program VDAT.exe}
\usage{
get_local_vdat_schema(vdat_exe_path = NULL)
}
\arguments{
\item{vdat_exe_path}{The full path to \code{VDAT.exe}. If \code{NULL}
(default) then the path to VDAT.exe must be in the PATH environment variable
of the system. See \code{\link{check_vdat}}.}
}
\value{
Schema (template) of VDAT CSV produced by installed version of VDAT.exe.
}
\description{
Get schema from local installation of Innovsea program VDAT.exe
}
\examples{
\dontrun{

#use if VDAT.exe in Windows system PATH variable
get_local_vdat_schema()

#or specify path to VDAT.exe
get_local_vdat_schema(vdat_exe_path = 
                     "C:/Program Files/Innovasea/Fathom/VDAT.exe")

}

}
