% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-vue.r
\name{get_local_vue_version}
\alias{get_local_vue_version}
\title{Get version of local installation of Innovasea program VUE.exe}
\usage{
get_local_vue_version(vue_exe_path = NULL)
}
\arguments{
\item{vue_exe_path}{The full path to \code{VUE.exe}. If \code{NULL}
(default) then the path to VUE.exe must be in the PATH environment variable
of the system. See \code{\link{check_vue}}.}
}
\value{
A list with \code{version} (version number) and \code{long_version} (full
string returned by VUE.exe).
}
\description{
Get version of local installation of Innovasea program VUE.exe
}
\examples{
\dontrun{

# use if VUE.exe in Windows system PATH variable
get_local_vue_version()

# or specify path to VUE.exe
get_local_vue_version(
  vue_exe_path =
    "C:/Program Files (x86)/Vemco/VUE"
)
}

}
