% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-install_ffmpeg.r
\name{get_os}
\alias{get_os}
\title{install ffmpeg}
\usage{
get_os()
}
\value{
value
}
\description{
install ffmpeg
}
\details{
checks for and installs ffmpeg into package directory for making animations
}
\author{
Todd Hayden
}
