% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filt-getMinLag.r
\name{getMinLag}
\alias{getMinLag}
\title{Calculate 'min_lag' for identifying potential false positive detections}
\usage{
getMinLag(detections, type = "GLATOS", detColNames = list())
}
\arguments{
\item{detections}{A data frame containing detection data with at least 3
columns containing 'transmitter', 'receiver', and 'timestamp'. Column names
are specified by \code{type}.}

\item{type}{A character string that contains the type of data that is being
passed in, for example, "OTN", "GLATOS", or "sample". Default is "GLATOS".}

\item{detColNames}{An optional list that contains the user-defined column 
names}
}
\value{
A data frame containing the data with a min_lag column appended
}
\description{
Calculate 'min_lag' column of a dataset, add this column into the dataset, 
and return the dataset with the new column
}
\details{
detColNames is defined as a list with the names of the required
  columns in \code{detections}, defined by \code{type}: \itemize{ \item
  \code{transmitters} is a character string with the name of the column 
  containing the ids of the transmitters ('transmission_id' for GLATOS data,
  'tagname' for OTN data, or 'transmitter' for sample data). \item
  \code{receivers} is a character string with the name of the column 
  containing the ids of the receivers ('receiver_sn' for GLATOS data,
  'receiver_group' for OTN data, or 'receiver' for sample data). \item
  \code{timestampCol} is a character string with the name of the column 
  containing datetime stamps for the detections (MUST be of class 'POSIXct') 
  ('detection_timestamp_utc' for GLATOS data, 'datecollected' for OTN data,
  or 'time' for sample data). }

Each value in the min_lag column defines the minimum number of
  seconds between either a time with the same transmitter and receiver if
  before or after the current time. This is calculated with the dplyr package
  in R. This min_lag column will be used in falseDetectionFilter.R to filter
  false data.

A new column (\code{min_lag}), indicating the minimum lag of each
entry (according to the definition above), is added to the input data frame.
}
\examples{
library(glatos)

#Example with GLATOS data
data(walleye_detections)

dtc2 <- getMinLag(walleye_detections)
  
}
\author{
A. Dini
}
