\name{glatos-package}
\Rdversion{1.2}
\alias{glatos-package}
\alias{glatos}
\docType{package}
\title{
Functions useful to members of the Great Lakes Acoustic Telemetry Observation System (www.data.glos.us/glatos); many more broadly relevant for simulation, processing, analysing, and visualizing acoustic telemetry data.
}
\description{
This package is in early stages of development. 
}
\details{
\tabular{ll}{
Package: \tab glatos\cr
Type: \tab Package\cr
Version: \tab 0.2\cr
Date: \tab 2016-03-17\cr
License: \tab GPL (2.0 or later)\cr
LazyLoad: \tab yes\cr
}
}
\author{
Chris Holbrook, Tom Binder, Todd Hayden

Maintainer: Chris Holbrook <cholbrook@usgs.gov>
}
\references{
www.data.glos.us/glatos
}

