% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-glatos_animals.r
\name{glatos_check_col_names}
\alias{glatos_check_col_names}
\alias{glatos_check_col_classes}
\title{Check column names and classes of a list or data.frame against requirements}
\usage{
glatos_check_col_names(x, req_cols)

glatos_check_col_classes(x, req_cols)
}
\arguments{
\item{x}{a data.frame, or object that inherits from data.frame, to check}

\item{req_cols}{a named list containing a character string with the class of
each required column; each element name is a required column name}
}
\description{
Check column names and classes of a list or data.frame against requirements

Check column classes of a list or data.frame against requirements
}
