% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-glatos_workbook.r
\name{glatos_workbook}
\alias{glatos_workbook}
\title{Constructor function for the class glatos_workbook}
\usage{
glatos_workbook(x)
}
\arguments{
\item{x}{A list containing data from a standard GLATOS data workbook
(*.xlsm) file.}
}
\value{
A list of class \code{glatos_workbook} created from a standard GLATOS
data workbook (*.xlsm) file with three elements:
\describe{
\item{metadata}{A list with data about the project.}
\item{animals}{A data frame with data about tagged animals.}
\item{receivers}{A data frame with data about receivers.}
}
}
\description{
Constructor function for the class glatos_workbook.
Currently barebones and only used inside read_glatos_workbook.
}
\note{
This function may be developed in the future to dictate conversion
constuction from a data frame.
}
\keyword{internal}
