% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{great_lakes_polygon}
\alias{great_lakes_polygon}
\title{An sf POLYGON object with Great Lakes coastline and some major 
tributaries.}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 4 rows and 9 columns.
}
\usage{
great_lakes_polygon
}
\description{
Created from \code{\link{greatLakesPoly}}. This is used as a
  default map background in several \link{glatos} functions.
}
\author{
Todd Hayden (coerced to sf by C. Holbrook)
}
\keyword{datasets}
