% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{greatLakesTrLayer}
\alias{greatLakesTrLayer}
\title{A TransitionLayer object that only allows transitions to occur within
water of the Great Lakes Basin.}
\format{
An object of class \code{TransitionLayer} of dimension 692 x 504 x 1.
}
\usage{
greatLakesTrLayer
}
\description{
A TransitionLayer object that only allows transitions to occur
within water (i.e., prohibits movement onto land). This dataset was
developed for non-linear interpolation of fish movement paths from telemetry
data and is used by default in \link{interpolate_path}.
}
\seealso{
\link{interpolate_path}, \link{gdistance}
}
\author{
Todd Hayden
}
\keyword{datasets}
