% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{higgins_lake_polygon}
\alias{higgins_lake_polygon}
\title{An sf POLYGON object with coastline of Higgins Lake}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 1 rows and 2 columns.
}
\usage{
higgins_lake_polygon
}
\description{
An sf POLYGON object with coastline of Higgins Lake, Michigan. 
Used as an example of a polygon representing a water body.
}
\author{
Chris Holbrook
}
\keyword{datasets}
