% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testdata.r
\name{higgins_lake_transition}
\alias{higgins_lake_transition}
\title{A transition object for Higgins Lake for testing make_transition}
\format{
A list comprised of a TransitionLayer and RasterLayer
(see \code{\link[=make_transition]{make_transition()}}).
}
\description{
A transition object, created from
\code{\link[=higgins_lake_polygon]{higgins_lake_polygon()}} for testing \code{\link[=make_transition]{make_transition()}}.
}
\section{Filename}{
 higgins_lake_transition.rds
}

\examples{
system.file("testdata", "higgins_lake_transition.rds", package = "glatos")

}
\author{
Chris Holbrook
}
