% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-impute_detections.r
\name{impute_detections}
\alias{impute_detections}
\title{Impute missed detections in a time series}
\usage{
impute_detections(dtc, dtc_col_names = list(timestampCol =
  "detection_timestamp_utc", transmitterCol = "transmitter", receiverCol =
  "receiver"), trnsm = NULL, rcvr = NULL, int = NULL)
}
\arguments{
\item{dtc_col_names}{A list with name of required columns for timestamp, transmitter, and receiver}

\item{trnsm}{Name of specific transmitter ids (or vector of ids) to subset; all others will be ignored and omitted from result}

\item{rcvr}{Name of specific receiver id (or vector of ids) to subset; all others will be ignored and omitted from result}

\item{int}{2-element vector of timestamps (POSIXct) that define the start and end of the time period of interest; if NA (default) then first and last timestamp in x will be used}

\item{x}{A dataframe containing (minimally) timestamp (POSIXct), transmitter id, and receiver id}
}
\value{
a dataframe with new column \code{detected} and imputed (missed) detections added (new rows)
}
\description{
Impute missed detections in a time series
}
\details{
Only supports fixed delay (period) transmitters.
}
\examples{
mydelay <- 45 #fixed 45 sec delay
tx <-  as.POSIXct('1999-01-01 12:59:40', tz="GMT") + mydelay*(1:99) #simulate 100 delays
tx <- sample(tx, size=30) #randomly select 30
tx <- sort(tx) #sort increasing
df <- data.frame(tx=tx, tid=123, rid=999) #make data frame
myint <- as.POSIXct(c("1999-01-01 13:00:00", "1999-01-01 14:15:00"), tz="UTC")
df2 <- impute_detections(df, 
         dtc_col_names = list(timestampCol = "tx", 
                              transmitterCol = "tid", 
                              receiverCol = "rid"),
         int = myint)
}
