% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-install_ffmpeg.r
\name{install_ffmpeg}
\alias{install_ffmpeg}
\title{Install external FFmpeg software into `glatos` package}
\usage{
install_ffmpeg()
}
\value{
message is returned to console signalling successful installation
}
\description{
Downloads and installs external ffmpeg into `glatos` package
directory system to enable creation animated videos of fish
movements
}
\details{
\code{install_ffmpeg} determines operating system
  (Windows or Mac) and downloads the latest stable version of
  FFmpeg software.  FFmpeg software is installed within glatos
  package and may be accessed using \code{system.file}
}
\examples{

\dontrun{

# call function
install_ffmpeg()

# determine version of FFmpeg installed and if install was successful
fle <- system.file("bin", "ffmpeg.exe", package = "glatos")
system2(fle, "-version")

}
}
\author{
Todd Hayden
}
