% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summ-residence_index.r
\name{interval_count}
\alias{interval_count}
\title{The function below takes a detection events data frame and determines the 
number of time bins in which detections were observed and returns the 
cumulative time covered by all bins, in days. Interval (bin) size is 
determined by the 'time_interval_size' argument.}
\usage{
interval_count(detections, time_interval_size)
}
\arguments{
\item{detections}{- data frame from detection_events (condensed = TRUE)}

\item{time_interval_size}{time increment string as in seq.Date 'by' argument}
}
\description{
For each event (row in detection events data frame), the function 
sequences from first_detection to last_detection by time_interval_size, then 
counts the number of unique intervals.
}
