% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis-make_transition.r
\name{make_transition}
\alias{make_transition}
\title{Create transition layer from polygon shapefile}
\usage{
make_transition(in_file = ".", output = "out.tif", output_dir = ".",
  res = c(0.1, 0.1))
}
\arguments{
\item{in_file}{character, file path to polygon shapefile (with
extension of *.shp)}

\item{output}{character, name of output file for rasterized \code{in_file}}

\item{output_dir}{character, directory that output file will be written}

\item{res}{two element vector that specifies the x and y dimension
of output raster cells.  Units of res are same as input
shapefile.}
}
\value{
A list with two elements:
\describe{
   \item{transition}{a geo-corrected transition raster layer where land = 0
      and water=1
  (see \code{gdistance})}
   \item{rast}{rasterized input layer of class \code{raster}}}
  Additonally, rasterized version of input shapefile is written to computer
  at \code{output_dir}
}
\description{
Create transition layer for \code{interpolat_path} from polygon shapefile.
}
\details{
\code{make_transition} uses gdalUtils::gdal_rasterize to
  convert a polygon shapefile into transition layer (see
  \code{gdistance}) where raster cell values on land = 0 and water
  = 1. Output raster is in same projection as input shapefile and
  has same extents as input shapefile.  Function requires that
  gdal is working on computer.  To determine if gdal is installed
  on your computer, see gdalUtils::gdal_rasterize.

output transition layer is corrected for projection
  distortions using \code{gdistance::geoCorrection}.  Adjacent
  cells are connected by 16 directions and transition function
  returns 0 (land) for movements between land and water and 1 for
  all over-water movements.
}
\examples{

# path to polygon shapefile
poly <- system.file("extdata", "shoreline.zip", package = "glatos")
poly <- unzip(poly, exdir = tempdir())

# make_transition layer
tst <- make_transition(poly[grepl("*.shp", poly)], res = c(0.1, 0.1))

# plot raster layer
# notice land = 1, water = 0
raster::plot(tst$rast)

# increase resolution
tst1 <- make_transition(poly[grepl("*.shp", poly)], res = c(0.01, 0.01))
raster::plot(tst1$rast)

}
\author{
Todd Hayden, Tom Binder, Chris Holbrook
}
