% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{oceanPoly}
\alias{oceanPoly}
\title{A SpatialPolygonDataFrame of the world's oceans.}
\format{An object of class \code{SpatialPolygonsDataFrame} with 2 rows and 2 columns.}
\source{
http//www.naturalearthdata.com/download/110m/physical/ne_110m_ocean.zip
}
\usage{
data(oceanPoly)
}
\description{
A SpatialPolygonDataFrame of the world's oceans. This is 
used as a default map background in several 
\link{glatos} functions when \code{type = "OTN"}.
}
\keyword{datasets}
