% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extdata.r
\name{otn_aat_animals}
\alias{otn_aat_animals}
\title{Example animal data from the OTN ERDDAP}
\format{CSV}
\source{
Ryan Gosse, Ocean Tracking Network
}
\usage{
system.file("extdata", "otn_aat_animals.csv", package = "glatos")
}
\description{
An example animal data file from the OTN ERDDAP
}
\section{Filename}{
 otn_aat_animals.csv
}

