% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extdata.r
\name{otn_aat_receivers}
\alias{otn_aat_receivers}
\title{Example station data from the OTN ERDDAP}
\format{CSV}
\source{
Ryan Gosse, Ocean Tracking Network
}
\usage{
system.file("extdata", "otn_aat_receivers.csv", package = "glatos")
}
\description{
An example receiver station data file from the OTN ERDDAP
}
\section{Filename}{
 otn_aat_receivers.csv
}

