% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extdata.r
\name{otn_aat_tag_releases}
\alias{otn_aat_tag_releases}
\title{Example tag release data from the OTN ERDDAP}
\format{
CSV
}
\source{
Ryan Gosse, Ocean Tracking Network
}
\usage{
system.file("extdata", "otn_aat_tag_releases.csv", package = "glatos")
}
\description{
An example tag release data file from the OTN ERDDAP
}
\section{Filename}{
 otn_aat_tag_releases.csv
}

