% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/positionHeatMap.r
\name{positionHeatMap}
\alias{positionHeatMap}
\title{Position Heat Maps}
\usage{
positionHeatMap(positions, resolution = 10, interval = 3600,
  legendPos = "right", AbsOrRelFish = "absolute",
  AbsOrRelPos = "absolute", AbsOrRelInt = "absolute",
  folder = "PositionHeatMaps")
}
\arguments{
\item{positions}{A dataframe containing detection data with at least 4 
columns named c('ID', 'timestamp', 'lat', 'lon'). The ID column contains 
a an individual identifier. The 'timestamp' column contains the 
datetime stamps for the positions (MUST be of class 'POSIXct'). The 
columns 'lat' and 'lon' are the position latitudes and longitudes, 
respectively.}

\item{resolution}{A numeric value indicating the spatial resolution 
(in meters) of the grid system used to make the heat maps. Default is 10 m.}

\item{interval}{A numeric value indicating the duration of each time bin 
(in seconds) for use in calculating NumIntervals (i.e., a surrogate for 
amount of time spent in each cell of the grid). Default is 3600 seconds 
(1 hour).}

\item{legendPos}{A character string indicating whether the legend should 
be displayed on the right side or bottom of the png image. Valid arguments 
are c("right", "bottom"). Default is "right".}

\item{AbsOrRelFish}{A character string indicating whether number of 
unique fish will display as absolute value (i.e, the actual number of fish 
positioned in each grid) or as relative number (i.e., % of total fish 
positioned). Valid arguments are c("absolute", "relative"). Default is 
"absolute".}

\item{AbsOrRelPos}{A character string indicating whether number of 
positions will display as absolute value (i.e, the actual number of 
positions in each grid) or as relative number (i.e., mean number of 
positions per fish detected). Valid arguments are c("absolute", 
"relative"). Default is "absolute".}

\item{AbsOrRelInt}{A character string indicating whether NumIntervals 
will display as absolute value (i.e, the actual number of detection 
intervals in each grid) or as relative number (i.e., mean number of 
intervals per fish detected ). Valid arguments are c("absolute", 
"relative"). Default is "absolute".}

\item{folder}{A character string indicating the name of the output folder 
(placed in the working directory). This will also be the name of the kml 
file for viewing the heat maps in Google Earth. Default is 
"PositionsHeatMaps".}

\item{inputTZ}{A character string specifying the timezone of the timestamps 
in the positions dataframe (default = "GMT").}
}
\value{
3 png files (one for each of number of positions, number of fish, 
 and number of intervals).
1 kml file for viewing the heat maps in Google Earth.
3 CSV files containing summary data used to create the plots.
}
\description{
Function for creating heat maps to display the spatial distribution of 
 acoustic telemetry positions.
}
\details{
NumIntervals is the number of unique fish x interval 
combinations that occurred each grid cell. For example, in 4 hours there 
are a total of 4 1-h intervals. If fish 'A' was positioned in a single grid 
cell during 3 of the 4 intervals, than NumIntervals for that fish and grid 
combination is 3. Intervals are determined by applying the findInterval 
function (base R) to a sequence of timestamps (class: POSIXct) created 
using seq(from = min(positions[, timestamp]), to = 
min(positions[, timestamp]), by = interval), where interval is the 
user-assigned interval duration in seconds. NumIntervals is a more 
robust surrogate than NumPositions for relative time spent in each grid 
in cases where spatial or temporal variability in positioning probability 
are likely to significantly bias the distribution of positions in the array.

NumFish, NumPositions, and NumIntervals can all be displayed as 
absolute or relative, which are specified using the AbsOrRelXXX arguments. 
"absolute" is the actual value. "relative" is the absolute value divided 
by the total number of fish positioned. Units for plots: NumFish = number 
of unique fish (absolute) or % of total fish positioned (relative); 
NumPositions = number of positions (absolute) or mean number of positions 
per fish (relative); NumIntervals = number of unique fish x interval 
combinations (absolute) or mean number of unique fish x interval 
combinations per fish (relative).
}
\examples{
data(lamprey_tracks)
#rename columns to match input arguments
names(lamprey_tracks)[c(1,2,6,7)] <- c("ID","timestamp","lat","lon")
positionHeatMap(lamprey_tracks)

}
\author{
T. R. Binder
}
