% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-prepare_tag_sheet.R
\name{prepare_tag_sheet}
\alias{prepare_tag_sheet}
\title{Loads the OTN tagging metadata sheet to prepare it for use in 
\code{convert_otn_to_att}}
\usage{
prepare_tag_sheet(path, header_line = 5, sheet_name = 2)
}
\arguments{
\item{path}{the path to the tagging sheet}

\item{header_line}{what line the headers are on}

\item{sheet_name}{the sheet name or number containing the metadata}
}
\value{
a data.frame created from the excel file.
}
\description{
Loads the OTN tagging metadata sheet to prepare it for use in 
\code{convert_otn_to_att}
}
\details{
The function takes the path to the tagging sheet, what line to start
reading the headers from, and what sheet in the excel file to use. It converts column names
to be used by \code{convert_otn_to_att}.
}
\examples{

#--------------------------------------------------
# EXAMPLE #1 - loading from NSBS tagging

library(glatos)
tag_path <- system.file("extdata", "otn_nsbs_tag_metadata.xls",
                        package = "glatos")

tags <- prepare_tag_sheet(tag_path, 5, 2)

}
\author{
Ryan Gosse
}
