% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{range_detection}
\alias{range_detection}
\title{Detection range data set}
\format{A data frame with 58309 rows and 30 variables}
\source{
F. Zomer, T. Hayden
}
\usage{
range_detection
}
\description{
Sample detection range data set from Lake Superior.
}
\details{
Data from a stationary detection range test conducted in 2018. Data
  are in standard GLATOS detection export format and are intened to 
  accompany detecton range analysis vignette.
}
\keyword{datasets}
