% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extdata.r
\name{raw_lamprey_workbook}
\alias{raw_lamprey_workbook}
\title{Raw GLATOS Workbook from St. Marys River Sea Lamprey project}
\format{
A macro-enabled Microsoft Excel workbook file (*.xlsm) with six
worksheets:
\describe{
\item{project}{project code, principal investigator and contact}
\item{locations}{descriptions of receiver array locations}
\item{proposed}{proposed receiver deployment locations and dates}
\item{deployment}{receiver deployment data (what, where, when, how)}
\item{recovery}{receiver recovery data (what, where, when, how)}
\item{tagging}{animal collection, tagging, and recovery data}
}
}
\source{
\url{http://glatos.glos.us/home/project/SMRSL}
}
\description{
A completed GLATOS workbook from St. Marys River Sea Lamprey
project.
}
\section{Filename}{
 SMRSL_GLATOS_20140828.xlsm
}

\examples{
system.file("extdata", "SMRSL_GLATOS_20140828.xlsm", package="glatos")

}
\author{
Chris Holbrook
}
