% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-read_glatos_receiver_locations.r
\name{read_glatos_receiver_locations}
\alias{read_glatos_receiver_locations}
\title{Read data from a GLATOS receiver location file}
\usage{
read_glatos_receiver_locations(rec_file, version = NULL)
}
\arguments{
\item{rec_file}{A character string with path and name of receiver location
file in standard GLATOS format (*.csv). If only file name is given, then
the file must be located in the working directory.}

\item{version}{An optional character string with 
the glatos file version number. If NULL (default value) then version 
will be determined by evaluating file structure. The only allowed 
values are \code{NULL} and \code{"1.0"}. Any other values will trigger 
an error.}
}
\value{
A data.frame of class \code{glatos_receiver_locations}:
}
\description{
Read data from a GLATOS receiver location (csv) file and return a 
data.frame of class \code{glatos_receiver_locations}.
}
\details{
Data are loaded using the \code{fread} function in the
\code{data.table} package and timestamps are coerced to POSIXct using the
\code{fastPOSIXct} function in the \code{fasttime} package. All timestamps
must be in UTC timezone per GLATOS standard.
}
\examples{
#get path to example receiver_locations file
rec_file <- system.file("extdata", 
  "receiver_locations_2011.csv", package = "glatos")
rcv <- read_glatos_receiver_locations(rec_file)

}
\author{
C. Holbrook (cholbrook@usgs.gov)
}
