% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-read_glatos_receivers.r
\name{read_glatos_receivers}
\alias{read_glatos_receivers}
\title{Read data from a GLATOS receiver location file}
\usage{
read_glatos_receivers(rec_file, version = NULL)
}
\arguments{
\item{rec_file}{A character string with path and name of receiver location
file in standard GLATOS format (*.csv). If only file name is given, then
the file must be located in the working directory. File must be a standard
GLATOS file (e.g., \emph{GLATOS_receiverLocations_yyyymmdd_xxxxxx.csv}) obtained
from GLATOSWeb Data Portal \url{http://glatos.glos.us}.}

\item{version}{An optional character string with the GLATOS file version
number. If NULL (default value) then version will be determined by
evaluating file structure. The only allowed values currently are \code{NULL} and
\code{"1.0"}. Any other values will trigger an error.}
}
\value{
A data.frame of class \code{glatos_receivers}.
}
\description{
Read data from a standard GLATOS receiver location (csv) file and return a
data.frame of class \code{glatos_receivers}.
}
\details{
Data are loaded using \link[data.table:fread]{fread} and timestamps are
coerced to POSIXct using \link[fasttime:fastPOSIXct]{fastPOSIXct}. All
timestamps must be 'YYYY-MM-DD HH:MM' format and in UTC timezone per GLATOS
standard.
}
\examples{
#get path to example receiver_locations file
rec_file <- system.file("extdata",
  "sample_receivers.csv", package = "glatos")

#note that code above is needed to find the example file
#for real glatos data, use something like below
#rec_file <- "c:/path_to_file/GLATOS_receiverLocations_20150321_132242.csv"

rcv <- read_glatos_receivers(rec_file)

}
\author{
C. Holbrook (cholbrook@usgs.gov)
}
