% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-read_otn_detections.r
\name{read_otn_detections}
\alias{read_otn_detections}
\title{Read data from a OTN detection file}
\usage{
read_otn_detections(det_file, format = "new")
}
\arguments{
\item{det_file}{A character string with path and name of detection file in
OTN detection extract format (*.csv or *.parquet). If only file name is
given, then the file must be located in the working directory.}

\item{format}{Either 'new' or 'old', denoting whether or not the file being
loaded is a CSV predating OTN's parquet rollout (old) or not (new).}
}
\value{
A data.frame of class \code{glatos_detections} that includes OTN
columns that do not map directly to GLATOS columns.
}
\description{
Read data from a standard OTN detection (csv) file and return
a data.frame of class \code{glatos_detections}.
}
\details{
Data are loaded using \code{\link[data.table:fread]{data.table::fread()}} package and timestamps
are coerced to POSIXct using \code{\link[lubridate:fast_strptime]{lubridate::fast_strptime()}}. All
times must be in UTC timezone per GLATOS standard.

Column names are changed to match GLATOS standard columns when possible.
Otherwise, OTN columns and column names are retained.
}
\examples{
# get path to example detection file
det_file <- system.file("extdata", "blue_shark_detections.csv", package = "glatos")
det <- read_otn_detections(det_file)

}
\author{
A. Nunes, \email{anunes@dal.ca}
}
