% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-read_vue_detection_csv.r
\name{read_vue_detection_csv}
\alias{read_vue_detection_csv}
\title{Read detection data exported from Innovasea VUE software}
\usage{
read_vue_detection_csv(src, show_progress = FALSE)
}
\arguments{
\item{src}{A character string with path and name of a CSV file produced
containing detection data exported from Innovasea VUE software. If only
file name is given, then the file must be located in the working directory.}

\item{show_progress}{Optional argument passed to \code{data.table::fread}'s
\code{showProgress}.}
}
\value{
A data.frame of class \code{vue_detections}.
}
\description{
Read detection data exported from Innovasea VUE software
}
\details{
Reading is done via \code{\link[data.table]{fread}}.

All timestamp columns are assumed to be in UTC.
}
\examples{
csv_file <- system.file("extdata",
  "VR2W_109924_20110718_1.csv",
  package = "glatos"
)

vue_det <- read_vue_detection_csv(csv_file)

}
\author{
C. Holbrook (cholbrook@usgs.gov)
}
