% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-read_glatos_workbook.r
\name{read_workbook_project}
\alias{read_workbook_project}
\title{Read Project sheet from GLATOS workbook file}
\usage{
read_workbook_project(wb_file)
}
\arguments{
\item{wb_file}{A character string with path and name of workbook in standard
GLATOS format (.xlsm or .xlsx). If only file name is given, then the file
must be located in the working directory. See the GLATOSWeb Data
Portal \url{https://glatos.org} for file format definitions.}
}
\value{
A list with six elements:
\enumerate{
\item \strong{project_code:} GLATOS Project Code.
\item \strong{principle_investigator:} Name of Principle Investigator.
\item \strong{pi_email:} Email address of Principle Investigator.
\item \strong{source_file:} Name of input \code{wb_file}.
\item \strong{wb_version:} Version of GLATOS workbook file.
\item \strong{created:} Timestamp when source file metadata was changed
(from \code{file.info(wb_file)$ctime}).
}
}
\description{
Read Project sheet from GLATOS workbook file
}
\examples{

# Example 1: Version 1.3 (xlsm)

# get path to example GLATOS Data Workbook
wb_file <- system.file("extdata",
  "walleye_workbook.xlsm",
  package = "glatos"
)

read_workbook_project(wb_file)


# Example 2: Version 1.4 (xlsx)

wb2_file <- system.file("extdata",
  "walleye_workbook.xlsx",
  package = "glatos"
)

read_workbook_project(wb2_file)

}
