% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{recLoc_example}
\alias{recLoc_example}
\title{A data frame containing receiver location data}
\format{An object of class \code{data.frame} with 4902 rows and 23 columns.}
\usage{
recLoc_example
}
\description{
A data frame containing receiver location data from the GLATOS Project 
HECSL. This is a sample of receiver location data in the standard 
GLATOS export format.
}
\author{
Todd Hayden
}
\keyword{datasets}
