% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summ-receiver_efficiency.r
\name{REI}
\alias{REI}
\title{Calculates receiver efficiency index for each receiver station}
\usage{
REI(detections, deployments)
}
\arguments{
\item{detections}{A \code{glatos_detections} object (e.g., produced by \link{as_glatos_detections}) containing required columns \strong{plus} columns for station (unique name representing the detection location; character) and species name (common_name_e; character)

\emph{OR} a data frame with the following columns whose names and types are specified below:
\describe{
\item{\code{animal_id}}{Individual unique animal identifier; character.}
\item{\code{detection_timestamp_utc}}{Detection timestamp in UTC; MUST be of class POSIXct.}
\item{\code{deploy_lat}}{Receiver deployment latitude for detection in decimal degrees (NAD83); numeric.}
\item{\code{deploy_long}}{Receiver deployment longitude for detection in decimal degrees (NAD83); numeric.}
\item{\code{ins_serial_no}}{Receiver serial number; character.}
\item{\code{station}}{unique name representing the detection location; character.}
\item{\code{common_name_e}}{species identifier; character.}
}}

\item{receivers}{A \code{glatos_receivers} object (e.g., produced by \link{as_glatos_receivers}) containing the required columns \strong{plus} a column for station (grouping column representing location of detection; character)

\emph{OR} a data frame with the following columns whose names and types are specified below:
\describe{
\item{\code{station}}{unique name representing the detection location; character.}
\item{\code{deploy_date_time}}{Receiver deployment timestamp; MUST be of class POSIXct.}
\item{\code{recover_date_time}}{Receiver recovery timestamp; MUST be of class POSIXct.}
\item{\code{deploy_lat}}{Receiver deployment latitude in decimal degrees (NAD83); numeric.}
\item{\code{deploy_long}}{Receiver deployment longitude in decimal degrees (NAD83); numeric.}
\item{\code{ins_serial_no}}{Receiver serial number; character.}
}

\strong{NOTE} Values of \code{station} in \code{receivers} and \code{detections} object should be the same.  Only matching \code{stations} in \code{receivers} and \code{detections} will be returned in output!}
}
\value{
a data frame of receivers with deployment latitude and deployment longitude and receiver efficiency index
}
\description{
The receiver efficiency index is number between 0 and 1 indicating the amount of relative activity at each receiver compared to the entire set of receivers, regardless of positioning. The function takes a set detections and a deployment history of the receivers to create a context for the detections. Both the amount of unique tags and number of species are taken into consideration in the calculation.

See:
\emph{(Ellis, R., Flaherty-Walia, K., Collins, A., Bickford, J., Walters Burnsed, Lowerre-Barbieri S. 2018. Acoustic telemetry array evolution: from species- and project-specific designs to large-scale, multispecies, cooperative networks,
\url{https://doi.org/10.1016/j.fishres.2018.09.015})}

REI() takes two arguments and is calculated as:
\deqn{
REI = (Tr/Ta) x (Sr/Sa) x (DDr/DDa) x (Da/Dr)
}
\itemize{
\item{Tr = The number of tags detected on the receievr}
\item{Ta = The number of tags detected across all receivers}
\item{Sr = The number of species detected on the receiver}
\item{Sa = The number of species detected across all receivers}
\item{DDa = The number of unique days with detections across all receivers}
\item{DDr = The number of unique days with detections on the receiver}
\item{Da = The number of days the array was active}
\item{Dr = The number of days the receiver was active}
}
}
\examples{
# load exmple detections file
dtc_file <- system.file("extdata", "walleye_detections.csv", package = "glatos")
dtc <- read_glatos_detections(dtc_file)
# load example receiver deployment/recovery file
recs <- system.file("extdata", "sample_receivers.csv", package = "glatos")
recs <- read_glatos_receivers(recs)
REI(dtc, recs)
}
\author{
Alex Nunes  \email{anunes@dal.ca}
}
