% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summ-residence_index.r
\name{residence_index}
\alias{residence_index}
\title{Generate the residence index from a set of detections}
\usage{
residence_index(detections, calculation_method = "kessel")
}
\arguments{
\item{detections}{A data.frame from the \code{\link{detection_events}}
function.}

\item{calculation_method}{A character string with the calculation method
using one of the following: \code{kessel}, \code{timedelta}, \code{aggregate_with_overlap}, or \code{aggregate_no_overlap}}
}
\value{
A data.frame of days_detected, residency_index, location,
mean_latitude, mean_longitude
}
\description{
This residence index tool will take a compressed or uncompressed detection
file and caculate the residency index for each station/receiver in the
detections. The information passed to the function is what is used to
calculate the residence index, make sure you are only passing the data you
want taken into consideration for the residence index (i.e. species,
stations, tags, etc.).
}
\details{
The \strong{Kessel} method converts both the \code{first_detection} and 
\code{last_detection} columns into a 
date with no hours, minutes, or seconds. Next it creates a list of the
unique days where a detection was seen. The size of the list is returned
as the total number of days as an integer. This calculation is used to
determine the total number of distinct days (T) and the total number of
distinct days per station (S).
Possible rounding error may occur as a detection on 2016-01-01 23:59:59
and a detection on 2016-01-02 00:00:01 would be counted as two days
when it is really 2-3 seconds.

\deqn{ RI = S/T}
\deqn{ RI = Residence Index}
\deqn{S = Distinct number of days detected at the station}
\deqn{T = Distinct number of days detected anywhere on the array}

The \strong{Timedelta} calculation method determines the first detection and the
last detection of all detections. The time difference is then taken as
the values to be used in calculating the residence index. The timedelta
for each station is divided by the timedelta of the array to determine
the residence index.

\deqn{
RI = Delta S/Delta T}

\deqn{RI = Residence Index}

\deqn{Delta S = Last detection time at a station - First detection time at the station}

\deqn{Delta T = Last detection time on an array - First detection time on the array}

The \strong{Aggregate With Overlap} calculation method takes the length of time of each
detection and sums them together. A total is returned. The sum for each station
is then divided by the sum of the array to determine the residence index.

\deqn{
RI = AwOS/AwOT}

\deqn{RI = Residence Index}

\deqn{AwOS = Sum of length of time of each detection at the station}

\deqn{AwOT = Sum of length of time of each detection on the array}

The \strong{Aggregate No Overlap} calculation method takes the length of time of each
detection and sums them together. However, any overlap in time between one or
more detections is excluded from the sum. For example, if the first detection
is from \code{2016-01-01 01:02:43} to \code{2016-01-01 01:10:12} and the second
detection is from \code{2016-01-01 01:09:01 }to \code{2016-01-01 01:12:43}, then the
sum of those two detections would be 10 minutes. A total is returned once all
detections of been added without overlap. The sum for each station is then
divided by the sum of the array to determine the residence index.

\deqn{
RI = AnOS/AnOT}

\deqn{RI = Residence Index}

\deqn{AnOS = Sum of length of time of each detection at the station, excluding any overlap}

\deqn{AnOT = Sum of length of time of each detection on the array, excluding any overlap}
}
\examples{
#get path to example detection file
det_file <- system.file("extdata", "walleye_detections.csv",
                         package = "glatos")
det <- read_glatos_detections(det_file)
detection_events <- glatos::detection_events(det)
rik_data <- glatos::residence_index(detection_events, calculation_method = 'kessel')
rit_data <- glatos::residence_index(detection_events, calculation_method = 'timedelta')
riawo_data <- glatos::residence_index(detection_events, calculation_method = 'aggregate_with_overlap')
riano_data <- glatos::residence_index(detection_events, calculation_method = 'aggregate_no_overlap')

}
\references{
Kessel, S.T., Hussey, N.E., Crawford, R.E., Yurkowski, D.J., O'Neill, C.V.
and Fisk, A.T., 2016. Distinct patterns of Arctic cod (\emph{Boreogadus
saida}) presence and absence in a shallow high Arctic embayment, revealed
across open-water and ice-covered periods through acoustic telemetry. Polar
Biology, 39(6), pp.1057-1068.
\url{https://www.researchgate.net/publication/279269147}
}
\author{
A. Nunes, \email{anunes@dal.ca}
}
