% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ri-ri_plot.r
\name{ri_plot}
\alias{ri_plot}
\title{Uses plotly to place the caluclated residence index on a map.}
\usage{
ri_plot(df, title = "Residence Index")
}
\arguments{
\item{df}{A data.frame from the \code{residence_index} containing
days_detected, residency_index, location, mean_latitude,
mean_longitude}

\item{title}{A string for the title of the plot}
}
\value{
A plotly plot_geo object
}
\description{
Uses plotly to place the caluclated residence index on a map.
}
\details{
This function uses plotly to place the caluclated RI on a map.
}
