% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simutil-rotate_points.r
\name{rotate_points}
\alias{rotate_points}
\title{Rotate points in a 2-d plane}
\usage{
rotate_points(x, y, theta, focus)
}
\arguments{
\item{x}{A numeric vector of x coordinates; minimum of 2.}

\item{y}{A numeric vector of y coordinates; minimum of 2.}

\item{theta}{A numeric scalar with the angle of rotation in degrees;
positive is clockwise.}

\item{focus}{A numeric vector of x (first element) and y (second element)
coordinates for the point around which \code{x} and \code{y} will rotate.}
}
\value{
A two-column data frame containing:
\item{x}{x coordinates}
\item{y}{y coordinates}
}
\description{
Rotate points around a point in a 2-d plane
}
\details{
Points are shifted to be centered at the focus, then rotated using a
rotation matrix, then shifted back to original focus.
}
\note{
This function is called from \code{\link[=crw_in_polygon]{crw_in_polygon()}}
}
\examples{
x <- runif(10,0,10)
y <- runif(10,0,10)
plot(x,y,type="b",pch=20)
foo <- rotate_points(x, y, 20, c(5, 5))
points(foo$x,foo$y,type="b",pch=20,col="red")

}
\author{
C. Holbrook (cholbrook@usgs.gov)
}
