% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{sample_detection_efficiency}
\alias{sample_detection_efficiency}
\title{Detection Efficiency data set}
\format{
A data frame with 7 rows and 5 variables
}
\source{
F. Zomer, T. Hayden
}
\usage{
sample_detection_efficiency
}
\description{
Sample detection Efficiency data set from Lake Papineau, Quebec, Canada.
}
\details{
Data from .
}
\keyword{datasets}
