% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis-make_transition.r
\name{scale_meters_to_degrees}
\alias{scale_meters_to_degrees}
\title{Get degree-scale equivalent of meter-scale distance on a spatial object}
\usage{
scale_meters_to_degrees(x, sf, ref = "center", epsg = 3175)
}
\arguments{
\item{x}{Distance, in meters (or units of \code{epsg}), to be converted to degrees
along each dimension of \code{sf}. May be a two element vector (denoting
distance along x and y axes, respectively) or a single number (same
distance along both axes).}

\item{sf}{An \code{sf} spatial object.}

\item{ref}{Reference point for scale. If \code{center} (default), then returned
value is the distance (in degrees), along each dimension, at the center of
\code{sf} that is equivalent to \code{x}. Other possible values are \code{min} and \code{max}.}

\item{epsg}{Code denoting coordinate reference system in which calculations
are made. Must be a Cartesian projection.}
}
\value{
A two-element vector with distance in x and y (longitude and
latitude) dimensions, respectively.
}
\description{
Get degree-scale equivalent of meter-scale distance on an \code{sf} object.
}
\details{
A helper function to determine input resolution to
\code{\link[=make_transition]{make_transition()}}.
}
\examples{

# how many long, lat equal to 5000 m at center of great_lakes_polygon?
scale_meters_to_degrees(x = 5000,
                        sf = great_lakes_polygon)

# how many long, lat equal to 5000 m at top of great_lakes_polygon?
scale_meters_to_degrees(x = 5000,
                        sf = great_lakes_polygon,
                        ref = "max")

# how many long, lat equal to 5000 m at bottom of great_lakes_polygon?
scale_meters_to_degrees(x = 5000,
                        sf = great_lakes_polygon,
                        ref = "min")

}
