% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summ-summarize_by_fish.r
\name{summarize_by_fish}
\alias{summarize_by_fish}
\title{Summarize GLATOS detections by animal_id, glatos_array, or both.}
\usage{
summarize_by_fish(dtc, by_col = c("animal_id", "glatos_array"))
}
\arguments{
\item{dtc}{A GLATOS detection data frame.}

\item{by_col}{Character string or vector containing names of columns
in dtc to group the count by. Result is}
}
\value{
A data frame with columns listed in by_col and count of detections
  in dtc corresponding to each unique by_col combination.
}
\description{
Summarize GLATOS detections by animal_id, glatos_array, or both.
}
\details{
TBD
}
\examples{
#get path to example detection file
det_file <- system.file("extdata", 
  "walleye_detections.zip",package="glatos")
det_file <- unzip(det_file, "walleye_detections.csv")
det <- read_glatos_detections(det_file)
summarize_by_fish(det)
summarize_by_fish(det, by = "transmitter_id")
summarize_by_fish(det, by = c("release_location", "sex", "glatos_array"))

}
\author{
C. Holbrook (cholbrook@usgs.gov)
}
