% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.glatos_detection.r
\name{summary.glatos_detections}
\alias{summary.glatos_detections}
\title{Summarize detections by animal and location}
\usage{
summary.glatos_detections(det, location_col = "glatos_array",
  as_matrix = FALSE, location_vals = NULL)
}
\arguments{
\item{det}{A data frame containing detection data with two columns 
described below and one column containing a location grouping variable 
specified by \code{location_col}. 
The following two columns must appear in \code{det}: 
\itemize{
  \item \code{animal_id} A character string with the name of the column 
       containing the individual animal identifier.
 \item \code{detection_timestamp_utc} is a character string with the name 
 of the column containing datetime stamps for the detections (MUST be of 
 class 'POSIXct').
}}

\item{location_col}{A character string with the name of the column 
containing the location grouping variable (e.g., "glatos_array") in 
quotes.}

\item{as_matrix}{A logical indicating if output should be formatted as
a matrix (TRUE) or data frame (FALSE; default).}

\item{location_vals}{A character vector with names of all possible values
in \code{location_col}. If \code{NULL} then all unique values in 
\code{det[ , "location_col"]} will be used.}
}
\value{
\itemize{
  \item{If \code{as_matrix = FALSE} (default): A data frame containing four
  columns, the animal identifier, the total number of detections for each
  animal, the total number of locations at which each animal was detected,
  and a space-separated character string containing a list of all unique
  locations each fish was detected.}
  \item{if \code{as_matrix = TRUE}: A list of three data frames where
  the first column in each row contains \code{animal_id} and the remaining 
  columns contain data (number of detections, first detection timestamp, 
  and last detection timestamp, respectively).}
  }
}
\description{
Summarize first and last detection timestamps and number of detections 
of each animal at each location.
}
\details{
The list of unique locations is pasted as a space-separated string.
}
\examples{

#get path to example detection file
det_file <- system.file("extdata", "walleye_detections.csv",
                         package = "glatos")
det <- read_glatos_detections(det_file)

#basic summary
ds <- summary(det)

#matrix summary
dsm <- summary(det, as_matrix = TRUE)

#specify all possible glatos_arrays from receiver file
#get example receiver data

rec_file <- system.file("extdata", "sample_receivers.csv",
                         package = "glatos")
rec <- read_glatos_receivers(rec_file) 

dsm2 <- summary(det, as_matrix = TRUE, 
                location_vals = sort(unique(rec$glatos_array)))

}
\author{
T. R. Binder and C. Holbrook
}
