% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summ-residence_index.r
\name{total_diff_days}
\alias{total_diff_days}
\title{The function below determines the total days difference.}
\usage{
total_diff_days(detections)
}
\arguments{
\item{Detections}{- data frame pulled from the compressed detections CSV}
}
\description{
The difference is determined by the minimal first_detection of every detection and the maximum last_detection of every detection.
Both are converted into a datetime then subtracted to get a timedelta. The timedelta
is converted to seconds and divided by the number of seconds in a day (86400). The function
returns a floating point number of days (i.e. 503.76834).
}
