\name{ulfx2csv}
\Rdversion{1.1}
\alias{ulfx2csv}
\title{
Convert Vemco ULFX (Vemco XML) file to two CSV files (one with detection data, one with receiver event data)
}
\description{
Create two CSV files (one with detection data, one with receiver event data) from a ULFX (Vemco XML) text file  
}
\usage{
ulfx2sv(ulfxFile)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{ulfxFile}{
Path and file name to ulfx file
}

}
\details{
TBD}
\value{
}
\references{
None
}
\author{
C. Holbrook
}
\note{
}


\seealso{
vrl2csv, vrl2ulfx
}
\examples{
}

