% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{vdat_csv_schema}
\alias{vdat_csv_schema}
\title{A schema for Innovasea Fathom (VDAT) CSV files}
\format{
An object of class \code{list} of length 2.
}
\usage{
vdat_csv_schema
}
\description{
A schema for Innovasea Fathom (VDAT) CSV files, produced by 'vdat.exe'
(Fathom Connect) or \code{\link[=vdat_convert]{vdat_convert()}}.
}
\details{
A list of lists of data frames that define record types (e.g.,
\code{DIAG}, \code{DET}, \code{EVENT_OFFLOAD}), columm names (e.g. \verb{Device Time (UTC)},
\code{Time}), and data types (e.g., \code{character}, \code{POSIXct}) for
comma-separated-values text file containing data produced by Innovasea's
\code{vdat.exe} (packaged with Fathom Connect software).

This is used to enforce column names and data types in
\code{\link[=read_vdat_csv]{read_vdat_csv()}}.
}
\author{
C. Holbrook
}
\keyword{datasets}
