% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filt-velocity_filter.r
\name{velocity_filter}
\alias{velocity_filter}
\title{Veloctiy filter}
\usage{
velocity_filter(detections, maximum_velocity = 10,
  filter_criteria = "both", drop_lead_and_lag_rows = TRUE)
}
\arguments{
\item{detections}{A \code{glatos_detections} object (e.g., produced by
\link{read_glatos_detections} or \link{read_otn_detections}).}

\item{maximum_velocity}{A numerical value expressed in m/s (i.e. 10 would be 10 m/s).}

\item{filter_criteria}{A string indicating whether to apply the filter to the lead, 
lag, or btohe velocites using the respective strings \code{lead}, \code{lag}, or 
\code{both} (default).}

\item{drop_lead_and_lag_rows}{A boolean to either drop (\code{TRUE}) or keep (\code{FALSE}) the
extra columns needed to calculate the velocity.}
}
\description{
Identify any detections where the lead, lag, or both velocites are greater than a 
maximum allowable velocity
}
\details{
The function groups by unique animal and orders the detections chronologically.
It then proceeds to determine the lead and lag station, time difference in seconds,
and geolocation for each detection. Using the determined lead and lag values it uses
\code{geosphere} to calculate the lead and lag diffence and then the velocity. A filter
column is then determined to be a 1 or 0 dpending on if the calculated velocity is
greater than the \code{maximum_velocity}.
}
\examples{
#' #get path to example detection file
det_file <- system.file("extdata", "walleye_detections.csv",
                         package = "glatos")
det <- read_glatos_detections(det_file)

det <- velocity_filter(det)
head(det)


}
