% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extdata.r
\name{video images}
\alias{video images}
\title{Video frames of walleye movements in Lake Huron}
\format{Folder contains 100 sequentially labeled .png image files}
\source{
\url{http://glatos.glos.us/home/project/HECWL}
}
\usage{
system.file("extdata", "frames", package="glatos")
}
\description{
Sequential images of walleye movements in Lake Huron
  for testing functionality of ffmpeg function.
}
\section{Filename}{
 frames
}

\author{
Todd Hayden
}
