% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vrl2csv.r
\name{vrl2csv}
\alias{vrl2csv}
\title{Convert Vemco VRL file(s) to CSV format (detection data only)}
\usage{
vrl2csv(vrl, vueExePath = "C:\\\\Program Files (x86)\\\\Vemco\\\\VUE")
}
\arguments{
\item{vrl}{A character string or vector with names of VRL file(s) or a 
single directory containing VRL files.}

\item{vueExePath}{A character string with path to VUE.exe (default may not 
always be correct).}
}
\value{
A character vector with output directory and file name(s).
}
\description{
Convert detection data from a VEMCO VRL file(s) to comma-separated-values 
  (CSV) format by invoking a system command in VUE (> 2.06; courtesy of Tim 
  Stone, Vemco).
}
\details{
If \code{vrl} is a directory, then all VRL files in that directory will be 
converted to CSV. Otherwise, only those files specified in \code{vrl} will 
be converted. Each output CSV file will have same name as its source VRL 
file.
}
\note{
Receiver event data are not exported because that functionality was not 
supported by the VUE system command at time of writing.

To get the path to VUE.exe, right click on the icon, select "Properties", 
and then copy text in "Target" box.

To create a CSV for time-corrected VRL files, first time-correct each 
file using the VRL editor in VUE (under Tools menu). To speed up that 
process, uncheck the "Import" checkbox next to each filename, then run 
\code{vrl2csv} to create a CSV for each edited (e.g. time-corrected) VRL.
}
\examples{
#get path to example VRL in this package
myVRL <- system.file("extdata", "VR2W_109924_20110718_1.vrl",
 package="glatos")
vrl2csv(dirname(myVRL)) #directory input
vrl2csv(myVRL) #file name input

}
\author{
C. Holbrook (cholbrook@usgs.gov)
}
