% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{walleye_controlTable}
\alias{walleye_controlTable}
\title{An example control data frame that determine order of location groups on the 
y-axis of an \link{abacusPlot} or \link{eventPlot}}
\format{An object of class \code{data.frame} with 24 rows and 2 columns.}
\usage{
walleye_controlTable
}
\description{
An example control data frame that determine order of location groups on the 
y-axis of an \link{abacusPlot} or \link{eventPlot}. This file is an example 
provided by Todd Hayden for use with example data \link{walleye_detections} 
with the functions \link{abacusPlot} and \link{eventPlot}.
}
\seealso{
\link{abacusPlot}, \link{eventPlot}, \link{walleye_detections}
}
\author{
Todd Hayden
}
\keyword{datasets}
