% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/walleye_detections.r
\docType{data}
\name{walleye_detections}
\alias{walleye_detections}
\title{A data frame containing fish detection data}
\description{
A data frame containing fish detection data from the GLATOS Project 
HECSL. This is a sample of detection data in the standard 
GLATOS export format.
}
\author{
Todd Hayden
}
