% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-write_vdat_csv.r
\name{write_vdat_csv}
\alias{write_vdat_csv}
\title{Write a vdat_list object to disk in Innovasea Fathom VDAT CSV format}
\usage{
write_vdat_csv(
  vdat,
  record_types = NULL,
  out_file = NULL,
  output_format = "csv.fathom",
  include_empty = FALSE,
  export_settings = NULL
)
}
\arguments{
\item{vdat}{A \code{vdat_list} object; e.g., produced by
\code{\link{read_vdat_csv}}..}

\item{record_types}{An optional vector of character strings with names of
record types to include in output. E.g., "DET" for detection records.
Default (\code{NULL}) will write all record types present in input CSV
\code{vdat}.}

\item{out_file}{Character string with name of CSV file to be written. If
\code{NULL} (default), or if \code{out_file} only contains a path, then file
name will be derived from the data source file name using
\code{tail(vdat$DATA_SOURCE_FILE$`File Name`, 1)}.}

\item{output_format}{Character string with output format. Options are:
\code{"csv.fathom"} (default) writes a single CSV file (for each input file)
with multiple record types interleaved; \code{"csv.fathom.split"} writes a
folder (for each input file) containing a separate CSV for each record type.}

\item{include_empty}{Logical (default = \code{FALSE}). If \code{output_format
= "csv.fathom.split"}, should files be written for empty objects.}

\item{export_settings}{(NOT YET IMPLEMENTED). Placeholder for future
specification of other options available via Fathom Data Export app. (E.g.,
'Data Types to Include', 'Data Filter', 'Filename Suffix', 'Time Offset in
Hours', 'Split CSV by UTC Day'.)}
}
\value{
A character string with full path and file name to output file.
}
\description{
Write a vdat_list object to disk in Innovasea Fathom VDAT CSV format
}
\details{
Writing is done via \code{\link[data.table]{fwrite}}.
}
\examples{
\dontrun{

# Example 1. Read and write a single file

vrl_file <- system.file("extdata", "detection_files_raw",
  "VR2W_109924_20110718_1.vrl",
  package = "glatos"
)

temp_dir <- tempdir()

csv_file <- vdat_convert(vrl_file, out_dir = temp_dir)

# utils::browseURL(temp_dir)

# read all record types
vdat <- read_vdat_csv(csv_file)

# write to single file (output_format = "csv.fathom")
temp_file <- tempfile(fileext = ".csv")
write_vdat_csv(vdat, out_file = temp_file)

# write to multiple files (fathom split option)
temp_dir2 <- tempdir()
write_vdat_csv(vdat, out_file = temp_dir2, 
  output_format = "csv.fathom.split")
}

}
\author{
C. Holbrook (cholbrook@usgs.gov)
}
