% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis-abacus_plot.r
\name{abacus_plot}
\alias{abacus_plot}
\title{Plot detection locations of acoustic transmitters over time}
\usage{
abacus_plot(det, location_col = "glatos_array", locations = NULL,
  ylab = NA, pch = NA, out_file = NULL, outFile = NULL, ...)
}
\arguments{
\item{det}{A \code{glatos_detections} object (e.g., produced by
\link{read_glatos_detections}).

\emph{OR} A data frame containing detection data with at least two columns, 
one of which must be named 'detection_timestamp_utc', described below,  
and another column containing a location grouping variable,
whose name is specified by \code{location_col} (see below).

The following column must appear in \code{det}: 
\describe{
\item{\code{detection_timestamp_utc}}{Detection timestamps; MUST be of 
class POSIXct.}
}}

\item{location_col}{A character string indicating the column name in
\code{det} that will be used as the location grouping variable (e.g.
"glatos_array"), in quotes.}

\item{locations}{An optional vector containing a list of locations
\code{location_col} to show in the plot. Plot order corresponds to order
in the vector (from bottom up). Should correspond to values in
\code{location_col}, but can contain values that are not in the det
data frame (i.e., can use this option to plot locations fish were
not detected).}

\item{ylab}{A character string indicating the y-axis label that will appear 
on the figure (default will match \code{location_col}).}

\item{pch}{A numeric value corresponding to the symbol to be used for
plotting. See \url{https://www.statmethods.net/advgraphs/parameters.html}.}

\item{out_file}{An optional character string with the name (including 
extension) of output image file to be created. File extension
will determine type of file written. For example, \code{"abacus_plot.png"}
will write a png file to the working directory. If \code{NULL} (default)
then the plot will be printed to the default plot device.
Supported extensions: png, jpeg, bmp, and tiff.}

\item{outFile}{Deprecated. Use \code{out_file} instead.}

\item{...}{Other plotting arguments that pass to "plot" function (e.g., col,
lwd, type).}
}
\value{
An image to the default plot device or a file containing the 
image if \code{out_file} is specified.
}
\description{
Plot detection locations of acoustic transmitters over time.
}
\details{
NAs are not allowed in any of the two required columns.

The locations vector is used to control which locations will
  appear in the plot and in what order they will appear. If no locations
  vector is supplied, the function will plot only those locations that
  appear in the \code{det} data frame and the order of locations on the 
  y-axis will be alphebetical from top to bottom.

By default, the function does not distinguish detections from
  different transmitters and will therefore plot all transmitters the same
  color. If more than one fish is desired in a single plot, a vector of
  colors must be passed to the function using the 'col =' argument. The color
  vector must be the same length as the number of rows in the detections data
  frame or the colors will be recycled.

Plotting options (i.e., line width and color) can be changed using
  optional graphical parameters 
  \url{http://www.statmethods.net/advgraphs/parameters.html} that are passed
  to "segments" (see ?segments).
}
\examples{
#get path to example detection file
det_file <- system.file("extdata", "walleye_detections.csv",
 package = "glatos")
det <- read_glatos_detections(det_file)

#subset one transmitter
det2 <- det[det$animal_id == 153, ]
	
#plot without control table and main tile and change color to red
abacus_plot(det, locations=NULL, 
  main = "TagID: 32054", col = "red")
	
#example with locations specified
abacus_plot(det, locations=c("DRF", "DRL", "FMP", "MAU", "PRS", "RAR",
   "DRM", "FDT"), main = "TagID: 32054", col = "red")

#plot with custom y-axis label and lines connecting symbols
abacus_plot(det, main = "TagID: 32054", type = "l",
	 col = "red")
  
}
\author{
T. R. Binder, edited by A. Dini
}
