% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abacusPlot.r
\name{abacusPlot}
\alias{abacusPlot}
\title{Plot detection locations of acoustic transmitters over time}
\usage{
abacusPlot(detections, detColNames = list(locationCol = "glatos_array",
  timestampCol = "detection_timestamp_utc"), controlTable = NULL,
  plotTitle = "", Ylab = NA, outFile = "AbacusPlot.png", ...)
}
\arguments{
\item{detections}{A data frame containing at least two columns with names 
specified by \code{detColNames}. The 'location' column contains the 
locations (typically 'glatos_array' or 'station' for GLATOS data) that 
will be plotted on the y-axis. The 'timestamp' column contains the 
datetime stamps for the detections (MUST be of class 'POSIXct').}

\item{detColNames}{An optional list of character strings with names of
  required columns in \code{detections}:
\itemize{
  \item \code{locationCol} A character scalar with the name (in quotes) of the 
     column containing the location codes to be plotted on the y-axis. The 
     default value ("glatos_array") is consistent with GLATOS standard.
  \item \code{timestampCol}A character scalar with the name (in quotes) of the 
     column containing the timestamp data to be plotted on the x-axis.
     The default value ("detection_timestamp_utc") is is consistent with 
     GLATOS standard.
}}

\item{controlTable}{Optional dataframe with two columns, c('location', and 
'y_order). The 'location' column is a character vector of locations to be 
plotted on the y-axis and the name of the 'location' column must match the 
'location' colum in \code{detections} dataframe 
(set by \code{detColNames}). The 'y_order' column specifies what order the 
grouping variable will appear on the y-axis (y_order increases as you move 
away from the x-axis).}

\item{plotTitle}{An optional character scalar that will apear at the top of 
the plot. Default is no title.}

\item{Ylab}{A character scalar indicating the y-axis label that will appear 
on the figure (default will match \code{detColNames$locationCol}).}

\item{outFile}{An optional character scalar with the name of the png file 
created (including file extension; default = "AbacusPlot.png").}

\item{...}{Other plotting arguments that pass to "plot" function 
(e.g., col, lwd, type).}
}
\value{
A png file containing the abacus plot (default name "AbacusPlot.png")
is written to the working directory.
}
\description{
Plot detection locations of acoustic transmitters over time
}
\details{
NAs are not allowed in any of the three required columns of 
  \code{events}.

The control table is used to control which locations will appear in 
the plot and in what order they will appear. If no controlTable is 
supplied, the function will plot only those locations that appear in the 
\code{detections} data frame and the order of locations on the y-axis will  
correspond to the order in which each location appears in the data frame.

By default, the function does not distinguish detections from  
different transmitters and will therefore plot all transmitters the same  
color. If more than one fish is desired in a single plot, a vector of colors  
must be passed to the function using the 'col =' argument. The color vector  
must be the same length as the number of rows in the detections data frame.

#' Alternatively, plots for multiple individual fish can be created 
by looping through and creating a separate plot on subsetted detections data.
Plotting options (i.e., line width and color) can be changed using optional 
graphical parameters 
\url{http://www.statmethods.net/advgraphs/parameters.html} 
that are passed to "segments" (see ?segments).
}
\examples{
data("walleye_detections") #example data

head(walleye_detections)

#subset one transmitter
walleye_detections <-
	 walleye_detections[walleye_detections$transmitter_id == 32123, ]
	
#plot without control table
abacusPlot(walleye_detections, controlTable=NULL, plotTitle = "TagID: 32123",
	 outFile="AbacusPlot_tag32123.png", col = "red")
	
#get example control table
data("walleye_controlTable") #example dataset
walleye_controlTable

#plot with control table
abacusPlot(walleye_detections, controlTable=walleye_controlTable, 
	 plotTitle = "TagID: 32123", outFile="AbacusPlot_tag32123_control.png", 
	 col = "red")

#plot with custom y-axis label and lines connecting symbols
abacusPlot(walleye_detections, controlTable=walleye_controlTable, 
	 plotTitle = "TagID: 32123", Ylab="Location (GLATOS Array)",
	 outFile="AbacusPlot_tag32123_control.png", 
	 col = "red", type="o")

}
\author{
T. R. Binder
}
