% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis-adjust_playback_time.r
\name{adjust_playback_time}
\alias{adjust_playback_time}
\title{Modify playback time of video}
\usage{
adjust_playback_time(
  scale_factor = 1,
  input,
  output_dir = getwd(),
  output = "new.mp4",
  overwrite = FALSE,
  ffmpeg = NA,
  diagnostic_mode = FALSE
)
}
\arguments{
\item{scale_factor}{multiplicative factor changes duration of
video playback. See details.}

\item{input}{character, path to video file (any FFmpeg supported
file type- *.mp4, *.wmv, see \link{make_frames})}

\item{output_dir}{character, output directory, default is working
directory}

\item{output}{character, output file name}

\item{overwrite}{logical, default is \code{overwrite = TRUE}}

\item{ffmpeg}{A file path (characer) to FFmpeg executable. This
argument is only needed if ffmpeg is not added to your system
path. For Windows machines, path must point to 'ffmpeg.exe',
located in the bin subfolder within the ffmpeg folder.  For
example on Windows machines,
"C:\\Users\\Username\\Documents\\ffmpeg-3.4.1-win64-static\\bin\\ffmpeg.exe").
On Mac, path must point to 'ffmpeg' within the 'bin'
subfolder "/home/directory/Documents/bin/ffmpeg".  see \link{make_video}}

\item{diagnostic_mode}{Logical (default = FALSE). If true, return value
is a character vector with FFMPEG output.}
}
\value{
One video animation will be written to \code{output_dir}
}
\description{
Speed up or slow down playback of video
}
\details{
\code{adjust_playback_time} adjusts playback speed of
video.  \code{scale_factor} controls the magnitude of speed-up
or slow-down by modifying the presentation timestamp of each
video frame. For more information, see
\href{https://trac.ffmpeg.org/wiki}{How to speed up/slow down a video}.
Values of \code{scale_factor} < 1 speed up playback and >1 slow
down video. In addition to changing playback, function can
change output format by specifying a different file extension in
\code{output}.
}
\examples{

\dontrun{

# load example frames 
frames <- system.file("extdata", "frames", package = "glatos")

# make video animation

make_video(input_dir = frames, pattern = "\%02d.png", output = "animation_av.mp4")




# slow video down by a factor of 10 
path <- file.path(getwd(), "animation.mp4")
adjust_playback_time(scale_factor = 10, input = path)

# slow video down by a factor of 10 and change format of output video
adjust_playback_time(scale_factor = 10, input = path, output = "slow.wmv")

# speed up video 
adjust_playback_time(scale_factor=0.5, input = path, output = "faster.mp4")
}
}
\author{
Todd Hayden, Tom Binder, Chris Holbrook
}
