% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-read_glatos_workbook.r
\name{cast}
\alias{cast}
\title{Cast a list of scalars to a new class}
\usage{
cast(
  x,
  new_class,
  old_class = c("logical", "character", "numeric", "Date", "POSIXct"),
  defer_exceptions = TRUE,
  ...
)
}
\arguments{
\item{x}{A list of scalars.}

\item{new_class}{A text string with name of new class.}

\item{old_class}{A character vector with names of classes in \code{x} that will
be cast to \code{new_class}. Any record with a different class will result in NA
(with error). Default value is \code{c("logical", "numeric", "character", "Date", "POSIXct")}.}

\item{defer_exceptions}{If TRUE (default value) then errors and warnings will
be returned as attributes with prefix "error_" or "warning_".}

\item{...}{Other arguments passed to the casting function (e.g., \code{tz = "US/Eastern"} when \code{new_class} is \code{POSIXct}).}
}
\value{
A vector of length same as \code{x} and class as \code{new_class}.
}
\description{
Cast a list of scalars, with potentially mixed classes, to a new class.
}
\details{
Written specifically for \code{readxl::read_excel} with \code{col_types = "list"} to
evaluate class of each record/row independently and then present user with
a single report of all errors (instead of sequential one. at. a. time).
}
\examples{

x <- list(TRUE, "A", NA, 3.1415, Sys.time(), Sys.Date(), "1997-05-13 12:43:21")

sapply(x, class)

cast(x, "character")

cast(x, "numeric")

cast(x, "Date")

cast(x, "POSIXct")

cast(x, "POSIXct", tz = "US/Pacific")

# empty list input returns empty result
cast(list(), "character")

# separate tz for each element
cast(x, "POSIXct", tz = c("US/Eastern", rep("US/Pacific", 5)))

# Only cast from if class is character
cast(x, "POSIXct", old_class = "character")

# Only cast from if class is character or POSIXct
cast(x, "character", old_class = c("character", "POSIXct"))

\dontrun{

# Bad (unsupported) new_class
cast(x, "foo")

# Bad (unsupported) old_class
cast(x, "character", old_class = c("character", "foo"))
}

}
