% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-read_glatos_workbook.r
\name{check_timezone}
\alias{check_timezone}
\title{Check a time zone string against OlsonNames()}
\usage{
check_timezone(tz, ignore.case = FALSE)
}
\arguments{
\item{tz}{a character string. The time zone specification to be used for the
conversion. Only values in \code{\link[=OlsonNames]{OlsonNames()}} are allowed. \code{""} is not allowed.}

\item{ignore.case}{logical. if FALSE, the pattern matching is case sensitive
and if TRUE (default), case is ignored during matching. Passed to \code{\link[=grep]{grep()}}.}
}
\value{
If \code{tz} is valid (depends on \code{ignore.case}), then it is returned.
Otherwise, \code{NA} is returned.
}
\description{
Check a time zone string against OlsonNames()
}
\examples{

x <- c("UTC", "US/Eastern", "US/EASTERN", "foo")

check_timezone(tz = x)

check_timezone(tz = x, ignore.case = TRUE)

}
