% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-vdat.r
\name{check_vdat}
\alias{check_vdat}
\title{Check path to Innovasea program VDAT.exe}
\usage{
check_vdat(vdat_exe_path = NULL)
}
\arguments{
\item{vdat_exe_path}{The full path to \code{VDAT.exe}. If \code{NULL}
(default) then the path to VDAT.exe must be in the PATH environment variable
of the system.}
}
\value{
Character string with command for calling VDAT.exe via
\code{system2}'s \code{\link{command}} argument.
}
\description{
Check path to Innovasea program VDAT.exe
}
\examples{

\dontrun{

#use Windows system PATH variable
check_vdat()


#use path to directory containing VDAT.exe
check_vdat(vdat_exe_path = "C:/Program Files/Innovasea/Fathom")


#use full path to VDAT.exe
check_vdat(vdat_exe_path = "C:/Program Files/Innovasea/Fathom/VDAT.exe")


}

}
