% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condense_events.R
\name{condense_events}
\alias{condense_events}
\title{Classify detection data into events}
\usage{
condense_events(detections, detColNames = list(locationCol = "glatos_array",
  animalCol = "animal_id", timestampCol = "detection_timestamp_utc"),
  timeSep = Inf)
}
\arguments{
\item{detections}{A data frame containing detection data with at least 
3 columns containing 'location', 'animal', and 'timestamp' columns. Column names are specified with \code{detColNames}.}

\item{detColNames}{A list with names of required columns in 
  \code{detections}: 
\itemize{
  \item \code{locationCol} is a character string with the name of the column 
       containing locations you wish to filter to (typically 'glatos_array' or 
         'station' for GLATOS data).
  \item \code{animalCol} is a character string with the name of the column 
         containing the individual animal identifier.
 \item \code{timestampCol} is a character string with the name of the column 
         containing datetime stamps for the detections (MUST be of class 
    'POSIXct').
}}

\item{timeSep}{Amount of time (in seconds) that must pass between 
sequential detections on the same receiver (or group of receivers, 
depending on specified location) before that detection is considered to 
belong to a new detection event. The default value \code{Inf}, will not 
define events based on elapsed time (only when location changes).}
}
\value{
Input dataframe with columns containing flags:
 \item{arrival}{Flag (0 or 1) representing the first detection in each event.} 
 \item{depart}{Flag (0 or 1) representing the last detection in each event.} 
 \item{event}{integer representing the event number}
 \item{TimeDiff}{Lagged time difference in seconds between successive detections.}
 \item{newfsh}{Flag representing first detection of each fish}
}
\description{
Reduce detection data from an acoustic telemetry receiver into discrete 
  detection events, defined by movement between receivers (or receiver 
  groups, depending on location), or sequential detections at the same 
  location that are separated by a user-defined threshold period of time.
}
\examples{
data("walleye_detections")
dtx <- condense_events(walleye_detections, timeSep = 3600)
}
\author{
T. R. Binder, Todd Hayden
}
