% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-convert_otn_to_att.r
\name{convert_otn_to_att}
\alias{convert_otn_to_att}
\title{Convert detections, tagging metadata, and deployment metadata to a format that ATT
accepts.}
\usage{
convert_otn_to_att(
  detectionObj,
  taggingSheet,
  deploymentObj = NULL,
  deploymentSheet = NULL,
  timeFilter = TRUE,
  crs = sp::CRS("+init=epsg:4326")
)
}
\arguments{
\item{detectionObj}{a data frame from \code{read_otn_detections}}

\item{taggingSheet}{a data frame from \code{prepare_tag_sheet}}

\item{deploymentObj}{a data frame from \code{read_otn_deployments}}

\item{deploymentSheet}{a data frame from \code{prepare_deploy_sheet}}

\item{timeFilter}{Whether the data should be filtered using the deployment and recovery/last download times of receivers.
Defaults to TRUE, if not all receiver metadata is available, this should be set to FALSE otherwise there will be data loss.}

\item{crs}{a \code{\link[=CRS-class]{sp::CRS}} object with geographic coordinate system for all spatial information (latitude/longitude). If none provided or \code{crs} is not recognized, defaults to WGS84.}
}
\value{
a list of 3 tibble::tibbles containing tag dectections, tag metadata,
  and station metadata, to be ingested by VTrack/ATT
}
\description{
Convert \code{glatos_detections}, OTN tagging metadata and OTN deployment metadata
to \code{ATT} format for use in the Animal Tracking Toolbox 
(\url{https://github.com/vinayudyawer/ATT}).
}
\details{
This function takes 3 data frames containing detections, tagging metadata, and
 deployment metadata from either \code{read_otn_deployments} or \code{prepare_deploy_sheet} 
 and transforms them into 3 \code{tibble::tibble} objects inside of a list. The input that 
 AAT uses to get this data product is located here:
 https://github.com/vinayudyawer/ATT/blob/master/README.md and our mappings
 are found here: https://gitlab.oceantrack.org/GreatLakes/glatos/issues/83
 in a comment by Ryan Gosse.
}
\examples{

#--------------------------------------------------
# EXAMPLE #1 - loading from Deployment Object

library(glatos)

dets_path <- system.file("extdata", "blue_shark_detections.csv", 
                         package = "glatos")
deploy_path <- system.file("extdata", "hfx_deployments.csv",
                           package = "glatos")
tag_path <- system.file("extdata", "otn_nsbs_tag_metadata.xls",
                        package = "glatos")

dets <- read_otn_detections(dets_path)
tags <- prepare_tag_sheet(tag_path, 5, 2)
deploy <- read_otn_deployments(deploy_path)

ATTdata <- convert_otn_to_att(dets, tags, deploymentObj = deploy)

#--------------------------------------------------
# EXAMPLE #2 - loading from Deployment Sheet

library(glatos)

dets_path <- system.file("extdata", "blue_shark_detections.csv", 
                         package = "glatos")
deploy_path <- system.file("extdata", "hfx_deploy_simplified.xlsx",
                           package = "glatos")
tag_path <- system.file("extdata", "otn_nsbs_tag_metadata.xls",
                        package = "glatos")

dets <- read_otn_detections(dets_path)
tags <- prepare_tag_sheet(tag_path, 5, 2)
deploy <- prepare_deploy_sheet(deploy_path)

ATTdata <- convert_otn_to_att(dets, tags, deploymentSheet = deploy)

}
\author{
Ryan Gosse
}
