% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crw.r
\name{crw}
\alias{crw}
\title{Simulate a correlated random walk}
\usage{
crw(theta = c(0, 5), stepLen = 10, initPos = c(0, 0), initHeading = 0,
  nsteps = 10000)
}
\arguments{
\item{theta}{A 2-element numeric vector with turn angle parameters 
(theta[1] = mean; theta[2] = sd) from normal distribution.}

\item{stepLen}{A numeric scalar with total distance moved in each step.}

\item{initPos}{A 2-element numeric vector with nital position (initPos[1]=x, 
initPos[2]=y).}

\item{initHeading}{A numeric scalar with initial heading in degrees.}

\item{nsteps}{A numeric scalar with number of steps to simulate.}
}
\value{
A two-column data frame containing:
\item{x}{x coordinates}
\item{y}{y coordinates}
}
\description{
Simulate a random walk as series of equal-length steps with turning angles 
  drawn from a normal distribution.
}
\details{
First, nsteps turn angles are drawn from a normal distribution. Second, the 
cumulative sum of the vector of turn angles defines the heading within each 
step. The x and y component vectors in each are then calculated and summed 
to obtain the simualted path.
}
\note{
Adapted from code provided by Tom Binder.
}
\examples{
foo <- crw(theta=c(0,5), stepLen=10, initPos=c(0,0), initHeading=0, nsteps=10)
plot(foo,type="b",pch=20,cex=0.2,asp=c(1,1)) 

}
\author{
C. Holbrook (cholbrook@usgs.gov)
}
