% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis-detection_bubble_plot.r
\name{detection_bubble_plot}
\alias{detection_bubble_plot}
\title{Plot number of tagged animals or detections on a map}
\usage{
detection_bubble_plot(det, location_col = "glatos_array",
  receiver_locs = NULL, map = NULL, out_file = NULL,
  background_ylim = c(41.3, 49), background_xlim = c(-92.45, -75.87),
  symbol_radius = 1, col_grad = c("white", "red"))
}
\arguments{
\item{det}{A data frame containing detection data with at least 
  five columns; four columns must be named 'animal_id', 
  'detection_timestamp_utc', 'deploy_lat', and 'deploy_long' as described 
  below and a fifth column containing a location grouping variable, is 
  specified using \code{location_col}.
  
  The following four columns must appear in \code{det}: 
\itemize{
  \item \code{animal_id} A character string with the name of the column 
         containing the individual animal identifier.
 \item \code{detection_timestamp_utc} is a character string with the name 
   of the column containing datetime stamps for the detections (MUST be of 
   class 'POSIXct').
 \item \code{deploy_lat} A character string with the name of the column
    containing latitude of the receiver in decimal degrees (NAD83).
 \item \code{deploy_long} A character string with the name of the column
    containing longitude of the receiver in decimal degrees (NAD83).
}}

\item{location_col}{A character string indicating the column name in the
detections data frame that will be used as the location grouping variable
(e.g. "glatos_array", "station").}

\item{receiver_locs}{An optional data frame containing at least 5 columns with 
receiver 'location', 'lat', 'lon', 'deploy_timestamp', and 
'recover_timestamp'.}

\item{map}{An optional SpatialPolygonsDataFrame or other
geo-referenced object to be plotted as the background for the plot. It is 
defined by \code{type}. If NULL, then the example Great Lakes polygon 
object (\code{data(greatLakesPoly)}) will be used.}

\item{out_file}{An optional character string with the name (including 
extension) of output file created. File extension will determine type of 
file written. For example, \code{"BubblePlot.png"} will write a png 
file to the working directory. If \code{NULL} (default) then the plot will 
be printed to the default plot device will be used. Supported extensions: 
png, jpeg, bmp, and tiff.}

\item{background_ylim}{two-element numeric vector that defines minimum and
maximum extents of the viewable plot area along the y-axis (i.e.,
longitude).}

\item{background_xlim}{A two-element numeric vector that defines minimum and
maximum extents of the viewable plot area along the x-axis (i.e.,
latitude).}

\item{symbol_radius}{Radius of each "bubble" on the plot in units of percent
of x-axis scale. Default value = 1 (i.e., 1 percent of x-axis).}

\item{col_grad}{A two-element character vector indicating the start and end
colors of the gradient scale used to color-code "bubbles".}
}
\value{
A data frame produced by 
  \code{glatos::summarize_detections(det, location_col = location_col, 
  receiver_locs = receiver_locs, summ_type = "location")}

If not out_file is specified, then an image is printed to the 
  default plot device. If out_file is specified, then an image of 
  specified type is written to \code{out_file}.
}
\description{
Make bubble plots showing the number of fish detected and number of 
  detections across a telemetry receiver network.
}
\details{
Data are summarized using \link{summarize_detections}.

If \code{receiver_locs} is specified (not NULL) then the plot will
  show all receivers in \code{receiver_locs} including those that detected
  none of the transmitters in \code{det}. Although this is helpful to view
  locations where fish were \emph{not} detections, it will also show and
  summarize data based on receivers that were not deployed during the time
  that a fish was at large. Therefore, \code{receiver_locs} only represent a
  subset of receivers present during the time interval most relevant to the
  data in \code{det}.

"col_grad" is used in a call to \link[grDevices]{colorRampPalette},
  which will accept a vector containing any two colors return by
  \link[grDevices]{colors} as character strings.
}
\examples{

#get path to example detection file
det_file <- system.file("extdata", "walleye_detections.csv",
  package = "glatos")
det <- read_glatos_detections(det_file)

#call with defaults
detection_bubble_plot(det)

#change symbol size and color
detection_bubble_plot(det, symbol_radius = 2, col_grad = c("grey90", "grey10"))

#Add all receivers

# get path to example receiver file
rec_file <- system.file("extdata", "sample_receivers.csv",
  package = "glatos")
rec <- read_glatos_receivers(rec_file)

detection_bubble_plot(det, receiver_locs = rec)


}
\seealso{
\code{\link{summarize_detections}}
}
\author{
T. R. Binder, edited by A. Dini
}
