% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filt-detection_events.r
\name{detection_events}
\alias{detection_events}
\title{Classify discrete events in detection data}
\usage{
detection_events(det, location_col = "glatos_array", time_sep = Inf,
  condense = TRUE)
}
\arguments{
\item{location_col}{A character string indicating the column name in the
detections data frame that will be used as the location grouping variable
(e.g. "glatos_array", "station").}

\item{time_sep}{Amount of time (in seconds) that must pass between 
sequential detections on the same receiver (or group of receivers, 
depending on specified location) before that detection is considered to 
belong to a new detection event. The default value \code{Inf}, will not 
define events based on elapsed time (only when location changes).}

\item{detections}{A data frame containing detection data with at least 
  five columns; four columns must be named 'animal_id', 
  'detection_timestamp_utc', 'deploy_lat', and 'deploy_long' as described 
  below and a fifth column containing a location grouping variable, is 
  specified using \code{location_col}.
  
  The following four columns must appear in \code{detections}: 
\itemize{
  \item \code{animal_id} A character string with the name of the column 
         containing the individual animal identifier.
 \item \code{detection_timestamp_utc} is a character string with the name 
   of the column containing datetime stamps for the detections (MUST be of 
   class 'POSIXct').
 \item \code{deploy_lat} A character string with the name of the column
    containing latitude of the receiver in decimal degrees (NAD83).
 \item \code{deploy_long} A character string with the name of the column
    containing longitude of the receiver in decimal degrees (NAD83).
}}
}
\value{
If \code{collapse = TRUE}, a data frame containing discrete 
  detection event data with the following columns:
\item{mean_latitude}{Mean latitude of detections comprising each event.}
	\item{mean_longitude}{Mean longitude of detections comprising each event.}
 \item{first_detection}{The time of the first detection in a given detection 
   event.} 
 \item{last_detection}{The time of the last detection in a given detection 
   event.}  
 \item{num_detections}{The total number of detection that comprised a given
	  detection event.}
 \item{res_time_sec}{The elapsed time in seconds between the first and last 
	detection in a given event.}
	
If \code{collapse = FALSE}, a data frame matching the input data frame 
\code{detections} with the following columns added:
 \item{time_diff}{Lagged time difference in seconds between successive 
   detections of each animal_id.}
 \item{arrive}{Flag (0 or 1) representing the first detection in each 
   event.} 
 \item{depart}{Flag (0 or 1) representing the last detection in each 
   event.} 
 \item{event}{Integer representing the event number.}
}
\description{
Reduce detection data from an acoustic telemetry receiver into discrete 
  detection events, defined by movement between receivers (or receiver 
  groups, depending on location), or sequential detections at the same 
  location that are separated by a user-defined threshold period of time.
}
\details{
mean_latitude and mean_longitude columns in the output dataframe are 
  the mean GPS locations for the detections comprising that detection event. 
  For example, if the a fish was detected at 3 receiver stations in a 
  glatos_array and glatos_array was selected as the location, then GPS 
  location for that event will be the mean of the latitude and longitude 
  for those three receiver stations (weighted based on the number of 
  detections that occurred on each station).
}
\examples{

#get path to example detection file
det_file <- system.file("extdata", "walleye_detections.csv",
                         package = "glatos")
det <- read_glatos_detections(det_file)

filt0 <- detection_events(det) #no time filter

#7-day filter
filt_7d <- detection_events(det , time_sep = 604800) 

}
\author{
T. R. Binder, T. A. Hayden, C. M. Holbrook
}
