% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detectionBubblePlot.r
\name{detectionBubblePlot}
\alias{detectionBubblePlot}
\title{Plot number of tagged animals or detections on a map}
\usage{
detectionBubblePlot(detections, receiverLocs = NULL, map = NULL,
  mapPars = list(xLimits = c(-94.5, -75), yLimits = c(41, 49.5), symbolRadius
  = 1, colGrad = c("white", "red"), showAll = FALSE),
  detColNames = list(locationCol = "glatos_array", animalCol = "animal_id",
  timestampCol = "detection_timestamp_utc", latCol = "deploy_lat", longCol =
  "deploy_long"), recColNames = list(locationCol = "glatos_array", latCol =
  "deploy_lat", longCol = "deploy_long", deploy_timestampCol =
  "deploy_date_time", recover_timestampCol = "recover_date_time"))
}
\arguments{
\item{detections}{A data frame containing detection data with at least 
5 columns containing 'location', 'animal', 'timestamp', 'latitude', 
and 'longitude' data. Default column names match GLATOS standard detection 
export file (e.g., '_detectionsWithLocs.csv'), but column names can also 
be specified with \code{detColNames}.}

\item{receiverLocs}{An optional data frame containing at least 5 columns with 
receiver 'location', 'lat', 'lon', 'deploy_timestamp', and 
'recover_timestamp'. Default column names match GLATOS standard receiver 
location file \cr(e.g., 'GLATOS_receiverLocations_yyyymmdd.csv'), but 
column names can also be specified with \code{recColNames}.}

\item{map}{An optional SpatialPolygonsDataFrame or other 
geo-referenced object to be plotted as the background for the plot. The 
default is a SpatialPolygonsDataFrame of the Great Lakes (e.g., 
\code{data(greatLakesPoly)}.}

\item{mapPars}{A list of optional mapping parameters (with exact names 
  matching below) including:
\itemize{
  \item \code{xLimits} is a two-element numeric vector that defines 
    minimum and maximum extents of the viewable plot area along the x-axis.
  \item \code{yLimits} is a two-element numeric vector that defines 
    minimum and maximum extents of the viewable plot area along the y-axis.
  \item \code{symbolRadius} is a numeric scalar that sets the radius of 
    each "bubble" on the plot in units of percent of x-axis scale. Default 
    value = 1 (i.e., 1% of x-axis).
  \item \code{colGrad} A two-element character vector indicating the start 
    and end colors of the gradient scale used to color-code "bubbles".
  \item \code{showAll} A logical (default = FALSE) indicating whether to 
    plot all receiver groups (TRUE) or only those receiver groups on 
    which the fish were detected (FALSE).
}}

\item{detColNames}{A list with names of required columns in 
  \code{detections}: 
  
\itemize{
  \item \code{locationCol} is a character string with the name of the column 
       containing the locations that will be plotted (typically 'glatos_array'  
    or 'station' for GLATOS standard detection export data).
  \item \code{animalCol} is a character string with the name of the column 
         containing the individual animal identifier (typically 'transmitter_id' 
    or 'animal_id' for GLATOS standard detection export data).
 \item \code{timestampCol} is a character string with the name of the column 
         containing datetime stamps for detections (MUST be of class 
    'POSIXct'; typically 'detection_timestamp_utc'for GLATOS standard 
    detection export data).  
 \item \code{latitudeCol} is a character string with the name of the column
    containing latitude of the receiver (typically 'deploy_lat' for 
    GLATOS standard detection export data). 
 \item \code{longitudeCol} is a character string with the name of the column
    containing longitude of the receiver (typically 'deploy_lat' for 
    GLATOS standard detection export data).
}}

\item{recColNames}{A list with names of required columns in 
  \code{receiverLocs}: 
\itemize{
  \item \code{locationCol} is a character string with the name of the column 
       containing the locations that will be plotted (typically 'glatos_array' 
    or 'station' for GLATOS standard detection export data).
 \item \code{latitudeCol} is a character string with the name of the column
    containing latitude of the receiver (typically 'deploy_lat' for 
    GLATOS standard detection export data). 
 \item \code{longitudeCol} is a character string with the name of the column
    containing longitude of the receiver (typically 'deploy_long' for 
    GLATOS standard detection export data).
 \item \code{deploy_timestampCol} is a character string with the name of 
    the column containing datetime stamps for receier deployments (MUST be 
    of class 'POSIXct'; typically 'deploy_date_time'for GLATOS standard 
    detection export data). 
 \item \code{recover_timestampCol} is a character string with the name of 
    the column containing datetime stamps for receier recover (MUST be of 
    class 'POSIXct'; typically 'recover_date_time'for GLATOS standard 
    detection export data). 
}}
}
\value{
A list containing the following data frames and columns: 
  \item{summaryNumFish}{
    \itemize{
		 	 \item \code{location}: user-specified 'location' labels
		 	 \item \code{Summary}: number of unique 'animals' detected at 'location'
      \item \code{meanLat}: mean latitude of receivers at 'location'
      \item \code{meanLon}: mean longitude of receivers at 'location'}}
  \item{summaryNumDetections}{
    \itemize{
		 	 \item \code{location}: user-specified 'location' labels
		 	 \item \code{Summary}: number of unique detections at 'location'
      \item \code{meanLat}: mean latitude of receivers at 'location'
      \item \code{meanLon}: mean longitude of receivers at 'location'}}

Two png files containing bubble plots for number of unique fish 
  detected ("BubblePlot_summaryNumFish.png") and total detections 
  ("BubblePlot_summaryNumDetections.png") are also written to the working 
  directory. Summary data for each plot are also written to CSV files 
  in the working directory.
}
\description{
Make bubble plots showing the number of fish detected and number of 
  detections across a telemetry receiver network.
}
\details{
If \code{mapPars$showAll} is TRUE then the plot will show all 
  receivers, including those that detected none of the transmitters in 
  \code{detections}. In that case, it will first be determined if a receiver 
  (or group of receivers, as defined by 'location') was in the water during 
  the time interval between the first and last detection in \code{detections}
  data frame. Receivers for which deployment-recovery intervals do not 
  with the time coverage of detections will not be included in the plot.

"ColGrad" is used in a call to \code{colorRampPalette()}, which 
  will accept a vector containing any two colors return by \code{colors()} 
  as character strings.
}
\examples{
#example detection data
data(walleye_detections) 
head(walleye_detections)

#call with defaults
detectionBubblePlot(walleye_detections)

#example receiver location data
data(recLoc_example) 
head(recLoc_example)

#view example map background
library(sp) #to avoid errors plotting SpatialPolygonsDataFrame
data(greatLakesPoly)
plot(greatLakesPoly)

detectionBubblePlot(walleye_detections, receiverLocs=recLoc_example,
  mapParms=list(symbolRadius = 1.4,colGrad = c("white", "blue"), showAll=T))

}
\author{
T. R. Binder
}
