% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-glatos-defunct.r
\name{glatos-defunct}
\alias{glatos-defunct}
\alias{check_dependencies}
\alias{make_video_ffmpeg}
\alias{install_ffmpeg}
\title{Defunct functions in glatos}
\description{
These functions are gone, no longer available.
}
\details{
\itemize{
\item \code{check_dependencies}: Removed in glatos 0.7.0.

\item \code{make_video_ffmpeg}: Removed in glatos 0.7.0.
Use \code{\link{make_video}} instead.

\item \code{install_ffmpeg}: Removed in glatos 0.7.0.

}
}
