% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-glatos-defunct.r
\name{glatos-defunct}
\alias{glatos-defunct}
\title{Defunct functions in glatos}
\description{
These functions are gone, no longer available.
}
\details{
\itemize{
\item \code{\link{check_dependencies}}:  This function is unneeded and defunct.

\item \code{\link{make_video_ffmpeg}}: This function has been replaced by \code{\link{make_video}}.

}
}
