% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-glatos-deprecated.r, R/util-vrl2csv.r
\name{glatos-deprecated}
\alias{glatos-deprecated}
\alias{vrl2csv}
\title{Deprecated functions in package \pkg{glatos}.}
\usage{
vrl2csv(
  vrl,
  outDir = NA,
  overwrite = TRUE,
  vueExePath = NA,
  showProgress = TRUE
)
}
\arguments{
\item{vrl}{A character string or vector with names of VRL file(s) or a
single directory containing VRL files.}

\item{outDir}{A character string directory where CSV files will be written.
If \code{NA} (default) then file(s) will be written to the current working
directory (e.g., \code{getwd()}).}

\item{overwrite}{Logical. If TRUE (default), output CSV file(s) will
overwrite existing CSV file(s) with same name in \code{outDir}. When FALSE,
'_n' (i.e., _1, _2, etc.) will be appended to names of output files that
already exist in \code{outDir}.}

\item{vueExePath}{An optional character string with directory containing
VUE.exe. If NA (default) then the path to VUE.exe must be added to the
PATH environment variable of your system. See Note below.}

\item{showProgress}{If TRUE (default) a progress bar and message are
displayed.}
}
\description{
The functions listed below are deprecated and will be defunct in
the near future. When possible, alternative functions with similar
functionality are also mentioned. Help pages for deprecated functions are
available at \code{help("<function>-deprecated")}.

These functions still work but will be removed (defunct) in the next version.

\itemize{
\item \code{vrl2_csv}: This function is deprecated, and will
be removed in the next version of this package.
Use \link{vue_convert} instead.
}
}
\section{\code{vrl2csv}}{

For \code{vrl2csv}, use \code{\link{vue_convert}}.
}

