% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-glatos-deprecated.r, R/util-vrl2csv.r
\name{glatos-deprecated}
\alias{glatos-deprecated}
\alias{vrl2csv}
\title{Deprecated functions in package \pkg{glatos}.}
\usage{
vrl2csv(
  vrl,
  outDir = NA,
  overwrite = TRUE,
  vueExePath = NA,
  showProgress = TRUE
)
}
\description{
The functions listed below are deprecated and will be defunct in
the near future. When possible, alternative functions with similar
functionality are also mentioned. Help pages for deprecated functions are
available at \code{help("<function>-deprecated")}.

These functions still work but will be removed (defunct) in the next version.

\itemize{
\item \code{vrl2_csv}: This function is deprecated, and will
be removed in the next version of this package.
Use \link{vue_convert} instead.
}
}
\section{\code{vrl2csv}}{

For \code{vrl2csv}, use \code{\link{vue_convert}}.
}

