% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-glatos.r
\docType{package}
\name{glatos}
\alias{glatos-package}
\alias{glatos}
\title{glatos: A package for the Great Lakes Acoustic Telemetry Observation System}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Functions useful to members of the Great Lakes Acoustic Telemetry Observation System https://glatos.org; many more broadly relevant to simulating, processing, analysing, and visualizing acoustic telemetry data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ocean-tracking-network/glatos}
  \item \url{https://ocean-tracking-network.github.io/glatos/}
  \item Report bugs at \url{https://github.com/ocean-tracking-network/glatos/issues}
}

}
\author{
\strong{Maintainer}: Christopher Holbrook \email{cholbrook@glfc.org}

Authors:
\itemize{
  \item Christopher Holbrook \email{cholbrook@glfc.org}
  \item Todd Hayden
  \item Thomas Binder
  \item Jon Pye
}

Other contributors:
\itemize{
  \item Mike O'Brien [contributor]
  \item Alex Nunes [contributor]
  \item Benjamin Hlina [contributor]
  \item Angela Dini [contributor]
  \item Ryan Gosse [contributor]
}

}
