% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-glatos_receivers.r
\name{glatos_receivers}
\alias{glatos_receivers}
\title{Constructor function for the class glatos_receivers}
\usage{
glatos_receivers(x)
}
\arguments{
\item{x}{A data.frame or data.table created from a standard glatos
receiver_location file.}
}
\value{
A data.frame of class \code{glatos_receivers}:
}
\description{
Constructor function for the class glatos_receivers.
Currently barebones and only used inside read_glatos_receivers.
}
\note{
This function may be developed in the future to dictate conversion
constuction from a data frame.
}
