% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greatLakesPoly.r
\docType{data}
\name{greatLakesPoly}
\alias{greatLakesPoly}
\title{A SpatialPolygonDataFrame with Great Lakes coastline and some major 
tributaries.}
\description{
A SpatialPolygonDataFrame with Great Lakes coastline and some major 
tributaries. This is used as a default map background in several 
\link{glatos} functions.
}
\author{
Todd Hayden
}
