% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{greatLakesPoly}
\alias{greatLakesPoly}
\title{A SpatialPolygonDataFrame with Great Lakes coastline and some major 
tributaries.}
\format{
An object of class \code{SpatialPolygonsDataFrame} with 4 rows and 8 columns.
}
\usage{
greatLakesPoly
}
\description{
A SpatialPolygonDataFrame with Great Lakes coastline and some 
major tributaries. This is used as a default map background in several 
\link{glatos} functions.
}
\author{
Todd Hayden
}
\keyword{datasets}
