% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greatLakesTrLayer.r
\docType{data}
\name{greatLakesTrLayer}
\alias{greatLakesTrLayer}
\title{A TransitionLayer object that only allows transitions to occur within water 
of the Great Lakes Basin.}
\description{
A TransitionLayer object that only allows transitions to occur within water
 (i.e., prohibits movement onto land). This dataset was developed for 
 non-linear interpolation of fish movement paths from telemetry data and 
 is used by default in \link{interpolatePath}.
}
\seealso{
\link{interpolatePath}, \link{gdistance}
}
\author{
Todd Hayden
}
