% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-kml_to_csv.r
\name{kml_to_csv}
\alias{kml_to_csv}
\title{KML To CSV Conversion}
\usage{
kml_to_csv(filePath, type = c("points", "lines", "polygons"))
}
\arguments{
\item{filePath}{The pathname for the kml file you wish to convert.}

\item{type}{Optional character string indicating the type(s) of feature(s)
to read from the kml file. Valid values are \code{c("points", "lines", and
  "polygons")}.}
}
\value{
A csv file (same name as input \code{filePath} but with \code{csv}
extension) is written to directory containing input \code{filePath}
with five columns \describe{
\item{name}{Feature name}
\item{feature_type}{Feature type}
\item{seq}{Sequential position in feature}
\item{longitude}{Longitude}
\item{latitude}{Latitude}
\item{altitude}{Altitude}
}
}
\description{
Function for extracting features (points, lines, polygons) from kml files and
writing them to csv files.
}
\details{
kmz files are not supported. Make sure exports from Google earth are
saved as kml. Or extract (unzip) kml from kmz.
}
\examples{

#Get example kml with two polygons
kml_file <- system.file("inst/extdata", "example_polygons.kml",
                        package = "glatos")
                        
kml_to_csv(kml_file)

}
