% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis-kml_workbook.r
\name{kml_workbook}
\alias{kml_workbook}
\title{Make a KML (or KMZ) file for viewing receiver and animal release 
  locations in Google Earth.}
\usage{
kml_workbook(zipFile, browse = F, kmz = F, labelSize = 0.6,
  iconSize = 0.6, showOngoing = T, endDate = "2020-01-01")
}
\arguments{
\item{browse}{A logical scalar. If TRUE, user is asked to select zipfile 
using windows explorer. Default value is FALSE.}

\item{kmz}{A logical scalar; If TRUE, a KMZ file (zipped KML file) will also 
be created. Default value is FALSE.}

\item{labelSize}{A numeric scalar with the size of placemark labels 
(only shown when placemark is highlighted by user).}

\item{iconSize}{A numeric scalar with the size of placemark icons.}

\item{showOngoing}{A logicalscalar that indicates if ongoing stations 
(missing recovery timestamp) should be included in result.}

\item{endDate}{End date (e.g. "YYYY-MM-DD") to be used for any ongoing 
stations (if showOngoing == T)}

\item{zipfile}{A character vector with the full path and filename of zipped 
GLATOS workbook (this is the **ZIPPED** archive that gets uploaded to 
GLATOSWeb)}
}
\value{
A KML (and optionally, KMZ) file, written to the directory that 
  contains the zipped GLATOS workbook. Nothing is returned to the R console.
}
\description{
Receiver data (deployment location, deployment timestamp, and 
recovery timestamp) and tagging data (release location, release timestamp) 
are imported from a zipped GLATOS Workbook archive and used to make a KML 
(and optionally, KMZ) for viewing receiver deployments  and release 
locations in Google Earth.
}
\details{
Receiver locations will be visible between deployment and recovery 
timestamps at each location. Release locations will be displayed when the 
display window includes the date of release.
}
\examples{
#TBD

}
\author{
C. Holbrook (cholbrook@usgs.gov)
}
