% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_transition.R
\name{make_transition}
\alias{make_transition}
\title{Create transition layer from polygon shapefile}
\usage{
make_transition(in_file, res = c(0.1, 0.1))
}
\arguments{
\item{in_file}{character, file path to polygon shapefile}

\item{res}{two element vector that specifies the x and y dimension
of output raster cells.  In example, units are degrees.}
}
\value{
returns geo-corrected transition raster layer where land = 0 and water =
  1 (see \code{gdistance})
}
\description{
Create transition layer for \code{interpolatePath} from polygon shapefile.
}
\details{
\code{make_transition} uses gdalUtils::gdal_rasterize to
  convert a polygon shapefile into transition layer (see
  \code{gdistance} where raster cell values on land = 0 and water
  = 1. Output raster is in same projection as input shapefile and
  has same extents as input shapefile.  Function requires that
  gdal is working on computer.  To determine if gdal is installed
  on your computer, see gdalUtils::gdal_rasterize.

output transition layer is corrected for projection
  distortions using \code{gdistance::geoCorrection}.  Adjacent
  cells are connected by 16 directions and transition function
  returns 0 (land) for movements between land and water and 1 for
  all over-water movements.
}
\examples{

# path to polygon shapefile
in_file <- system.file("extdata", "shoreline.zip", package = "glatos")
gl <- unzip(in_file, "shoreline.shp" )

# make_transition layer
tst <- make_transition(gl, res = c(0.1, 0.1))

}
\author{
Todd Hayden, Tom Binder, Chris Holbrook
}
