% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/min_lag.R
\name{min_lag}
\alias{min_lag}
\title{Minimum lag filter}
\usage{
min_lag(dtc)
}
\arguments{
\item{dtc}{An object of class \code{glatos_detections} or data
frame containing acoustic detection data with
\code{detection_timestamp}, \code{transmitter_id}, and
\code{receiver_sn}.  \code{min_lag} is loosely based on the the
"short interval" method described by Pincock (2012) and
replicates the \code{min_lag} column in the standare glatos
export file. In this case (GLATOS), \code{min_lat} is defined for each
detection as the shortest interval (in seconds) between either
the previous or next detection (whichever is closest) of the same
transmitter on the same receiver.}
}
\value{
a column is added to input dataframe with the shortest time interval
for each observation
}
\description{
Calculate minimum time interval between successive detections for identifying
false detections
}
\details{
Calculates minimum time interval between detections for each tag on
a receiver

A new column (\code{min_lag}) is added to the input
    dataframe that represents the time (in seconds) between the
    current detection and the next detection (either before or
    after) of the same transmitter on the same receiver. This
    function replicates the 'min_lag' column included in the
    standard glatos export.
}
\examples{

# load detection file

det_file <- system.file("extdata", "walleye_detections.csv",
                         package = "glatos")
dtc <- read_glatos_detections(det_file)

# calculate min_lag
min_lag(dtc)



}
\author{
Todd Hayden
}
