% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/min_lag.R
\name{min_lag}
\alias{min_lag}
\title{Minimum lag filter}
\usage{
min_lag(detections, detColNames = list(transmitter_id = "transmitter_id",
  timestampCol = "detection_timestamp_utc", receiver_sn = "receiver_sn"))
}
\arguments{
\item{detections}{A data frame containing detection
data (e.g., from the standard GLATOS detection export file
'*_detectionsWithLocs.csv').  Must contain the columns
\code{detection_timestamp}, \code{transmitter_id}, and
\code{receiver_sn} or equivalent (specified by
\code{detColNames}). \code{min_lag} is loosely based on the the
"short interval" method described by Pincock (2012) and replicates
the \code{min_lag} column in the standare glatos export file. In
this case (GLATOS), it is defined for each detection as the
shortest interval (in seconds) between either the previous or next
detection (whichever is closest) of the same transmitter on the
same receiver.}

\item{detColNames}{An optional list of character strings with names of required columns in
    \code{detections}:
\itemize{
   \item \code{transmitter_id} A character scalar with the name (in quotes) of the column
    containing transmitter_id.  The default value
    ('transmitter_id') is consistent with the GLATOS standard.

\item \code{timestampCol} A character scalar with the name (in
    quotes) of the column containing the timestamp data. The
    default value ("detection_timestamp_utc") is consistent with
    GLATOS standard.
\item \code{receiver_sn} A character scalar with the name (in quotes) of the column
    containing receiver_id.  The default value
    ('receiver_id') is consistent with the GLATOS standard.
}}
}
\value{
a column is added to input dataframe with the shortest time interval for each observation
}
\description{
Calculate minimum time interval between successive detections for identifying false detections
}
\details{
Calculates minimum time interval between detections for each tag on
a receiver

A new column (\code{min_lag}) is added to the input
    dataframe that represents the time (in seconds) between the
    current detection and the next detection (either before or
    after) of the same transmitter on the same receiver. This
    function replicates the 'min_lag' column included in the
    standard glatos export.
}
\examples{
data('walleye_detections') # example data

head(walleye_detections)

# calculate min_lag
dtx <- min_lag(walleye_detections)



}
\author{
Todd Hayden
}
