% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-prepare_deploy_sheet.r
\name{prepare_deploy_sheet}
\alias{prepare_deploy_sheet}
\title{Loads the OTN receiver deployment metadata sheet to prepare it for use in 
\code{convert_otn_to_att}}
\usage{
prepare_deploy_sheet(path, start = 1, sheet = 1)
}
\arguments{
\item{path}{the path to the deployment sheet}

\item{start}{what line to start reading data in from}

\item{sheet}{the sheet name or number containing the metadata}
}
\value{
a data.frame created from the excel file.
}
\description{
Loads the OTN receiver deployment metadata sheet to prepare it for use in 
\code{convert_otn_to_att}
}
\details{
The function takes the path to the deployment sheet, what line to start
reading from, and what sheet in the excel file to use. It converts column names
to be used by \code{convert_otn_to_att}.
}
\examples{

#--------------------------------------------------
# EXAMPLE #1 - loading from NSBS simplified Deployments

library(glatos)
deploy_path <- system.file("extdata", "hfx_deploy_simplified.xlsx",
                        package = "glatos")

deploy <- prepare_deploy_sheet(deploy_path, 5, 2)

}
\author{
Ryan Gosse
}
