% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extdata.r
\name{raw_walleye_detections}
\alias{raw_walleye_detections}
\title{Zipped GLATOS detection file from Huron Erie Corridor Walleye project}
\format{A zipped walleye detection file in detection file format 1.3:}
\source{
\url{http://glatos.glos.us/home/project/HECWL}
}
\usage{
system.file("extdata", "walleye_detections.zip", package="glatos")
}
\description{
An example detection file
}
\section{Filename}{
 walleye_detections.zip
}

\author{
Todd Hayden
}
