% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-read_glatos_detections.r
\name{read_glatos_detections}
\alias{read_glatos_detections}
\title{Read data from a GLATOS detection file}
\usage{
read_glatos_detections(det_file, version = NULL)
}
\arguments{
\item{det_file}{A character string with path and name of detection file in 
standard GLATOS format (*.csv). If only file name is given, then the 
file must be located in the working directory.}

\item{version}{An optional character string with 
the glatos file version number. If NULL (default value) then version 
will be determined by evaluating workbook structure. The only allowed 
values are \code{NULL} and \code{"1.3"}. Any other values will trigger 
an error.}
}
\value{
A data.frame of class \code{glatos_detections}:
}
\description{
Read data from a GLATOS detection (csv) file and return a data.frame of  
class \code{glatos_detections}.
}
\details{
Data are loaded using the \code{fread} function in the 
\code{data.table} package and timestamps are coerced to POSIXct usign the 
\code{fastPOSIXct} function in the \code{fasttime} package. All times must be
in UTC timezone per GLATOS standard.
}
\examples{
#get path to example detection file
det_file <- system.file("extdata", 
  "walleye_detections.zip",package="glatos")
det_file <- unzip(det_file, "walleye_detections.csv")
det <- read_glatos_detections(det_file)

}
\author{
C. Holbrook (cholbrook@usgs.gov)
}
