% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_glatos_workbook.r
\name{read_glatos_workbook}
\alias{read_glatos_workbook}
\title{Read data from a GLATOS workbook}
\usage{
read_glatos_workbook(workbook, wb_version = NULL, read_all = FALSE)
}
\arguments{
\item{workbook}{A character string with path and name of workbook in 
standard GLATOS format (*.xlsm). If only file name is given, then the 
file must be located in the working directory.}

\item{wb_version}{An optional character string with 
the workbook version number. If NULL (default value) then version will be
determined by evaluating workbook structure. The only allowed values 
are \code{NULL} and \code{"1.3"}. Any other values will trigger an error.}

\item{read_all}{If TRUE, then all columns and sheets
(e.g., user-created "project-specific" columns or sheets) in the workbook
will be imported. If FALSE (default value) then only columns in the 
standard GLATOS workbook will be imported (project-specific columns will 
be ignored.)}
}
\value{
A list of class \code{glatos_workbook} with six elements:
\describe{
  \item{metadata}{A list with data about the project.}
  \item{instruments}{A data frame with information about stations.}
  \item{animals}{A data frame with data about animals.}
}
}
\description{
Read data from a GLATOS workbook (xlsm file) and return a list of class 
\code{glatos_workbook}.
}
\details{
If \code{read_all = TRUE} then the type of data in each user-defined 
column will be 'guessed' by read_excel; this may throw some warnings.
}
\examples{
#get path to example GLATOS Data Workbook
wb_file <- system.file("extdata", 
  "SMRSL_GLATOS_20140828.xlsm",package="glatos")
wb <- read_glatos_workbook(wb_file)

}
\author{
C. Holbrook (cholbrook@usgs.gov)
}
