% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-read_otn_deployments.R
\name{read_otn_deployments}
\alias{read_otn_deployments}
\title{Read data from a OTN detection file}
\usage{
read_otn_deployments(deployment_file)
}
\arguments{
\item{deployment_file}{A character string with path and name of detection file in
OTN deploymeny format (*.csv). If only file name is given, then the
file must be located in the working directory.}
}
\value{
A data.frame of class \code{glatos_receivers} that includes OTN
columns that do not map directly to GLATOS columns.
}
\description{
Read data from a standard OTN detection (csv) file and return
a data.frame of class \code{glatos_detections}.
}
\details{
Data are loaded using \code{\link[data.table]{fread}} package and timestamps
are coerced to POSIXct using the \code{\link[fasttime]{fastPOSIXct}}. All
times must be in UTC timezone per GLATOS standard.

Column names are changed to match GLATOS standard columns when possible.
Otherwise, OTN columns and column names are retained.
}
\examples{
#get path to example deployments file
deployment_file <- system.file("extdata", "hfx_deployments.csv",
                         package = "glatos")
det <- read_otn_deployments(deployment_file)

}
\author{
A. Nunes, \email{anunes@dal.ca}
}
