% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recLoc_example.r
\docType{data}
\name{recLoc_example}
\alias{recLoc_example}
\title{A data frame containing receiver location data}
\description{
A data frame containing receiver location data from the GLATOS Project 
HECSL. This is a sample of receiver location data in the standard 
GLATOS export format.
}
\author{
Todd Hayden
}
