% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summ-receiver_efficiency.r
\name{REI}
\alias{REI}
\title{Calculates a returns a list of each station and the REI (defined here)}
\usage{
REI(detections, deployments)
}
\arguments{
\item{detections}{a glatos detections class data table}

\item{deployments}{a glatos receivers class data table}
}
\value{

}
\description{
The receiver efficiency index is number between 0 and 1 indicating the amount of relative activity at each receiver compared to the
entire set of receivers, regardless of positioning. The function takes a set detections and a deployment history of the receivers to
create a context for the detections. Both the amount of unique tags and number of species are taken into consideration in the calculation.

REI() takes two arguments. The first is a dataframe of detections the detection timstamp, the station identifier, the species, and the
tag identifier. The next is a dataframe of deployments for each station. The station name should match the stations in the detections.
The deployments need to include a deployment date and recovery date.

 \deqn{
REI = (Tr/Ta) x (Sr/Sa) / (DDa/DDr) / Dr

Tr = The number of tags detected on the receievr
Ta = The number of tags detected across all receivers
Sr = The number of species detected on the receiver
Sa = The number of species detected across all receivers
DDa = The number of unique days with detections across all receivers
DDr = The number of unique days with detections on the receiver
Dr = The number of days the receiver was active
}
}
\examples{


}
\author{
Alex Nunes  \email{anunes@dal.ca}
}
