% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summ-receiver_efficiency.r
\name{REI}
\alias{REI}
\title{Calculates a returns a list of each station and the REI (defined here)}
\usage{
REI(detections, deployments)
}
\arguments{
\item{detections}{a glatos detections class data table}

\item{deployments}{a glatos receivers class data table}
}
\value{

}
\description{
The receiver efficiency index is number between 0 and 1 indicating the amount of relative activity at each receiver compared to the
entire set of receivers, regardless of positioning. The function takes a set detections and a deployment history of the receivers to
create a context for the detections. Both the amount of unique tags and number of species are taken into consideration in the calculation.

\emph{(Ellis, R., Flaherty-Walia, K., Collins, A., Bickford, J., Walters Burnsed, Lowerre-Barbieri S. 2018. Acoustic telemetry array evolution: from species- and project-specific designs to large-scale, multispecies, cooperative networks, 
 \url{https://doi.org/10.1016/j.fishres.2018.09.015})}

REI() takes two arguments. The first is a dataframe of detections the detection timstamp, the station identifier, the species, and the
tag identifier. The next is a dataframe of deployments for each station. The station name should match the stations in the detections.
The deployments need to include a deployment date and recovery date.

 \deqn{
REI = (Tr/Ta) x (Sr/Sa) x (DDr/DDa) x (Da/Dr)
}
\itemize{
\item{Tr = The number of tags detected on the receievr}
\item{Ta = The number of tags detected across all receivers}
\item{Sr = The number of species detected on the receiver}
\item{Sa = The number of species detected across all receivers}
\item{DDa = The number of unique days with detections across all receivers}
\item{DDr = The number of unique days with detections on the receiver}
\item{Da = The number of days the array was active}
\item{Dr = The number of days the receiver was active}
}
}
\examples{
det_file <- system.file("extdata", "hfx_detections.csv",
package = "glatos")

dep_file <- system.file("extdata", "hfx_deployments.csv",
                        package = "glatos")

hfx_deployments <- glatos::read_otn_deployments(dep_file)
dets <- glatos::read_otn_detections(det_file)

hfx_receiver_efficiency_index <- glatos::REI(dets,hfx_deployments)


}
\author{
Alex Nunes  \email{anunes@dal.ca}
}
