% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{sample_detection_efficiency}
\alias{sample_detection_efficiency}
\title{Detection Efficiency data set}
\format{
A data frame with 7 rows and 5 variables
\describe{
\item{distance_m}{distance away from the receiver in meters}
\item{avg_percent}{average detection efficiency}
\item{std_dev}{standard deviation of detection efficiency}
\item{avg_percent_d}{average detection efficiency in decimal form needs
to be created by dividing \code{avg_percent} by 100}
\item{intercept}{y-intercept used for third order polynomial, set at 100.
Needs to be added to the original dataframe}
}
}
\source{
B.L. Hlina
}
\usage{
sample_detection_efficiency
}
\description{
Sample detection efficiency data set from Lake Papineau, Quebec, Canada.
}
\details{
Data is from a preliminary range test, where tags were deployed at
set distances away from a VR2W receiver for 24 hours. Once downloaded the vrl
files were used in Vemco's Range Testing Software to produced this dataset.
}
\keyword{datasets}
