% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extdata.r
\name{shoreline}
\alias{shoreline}
\title{zipped polygon shapefile of Great Lakes}
\format{
shapefile
}
\source{
\url{http://glatos.glos.us/home}
}
\description{
Polygon coastline of Great Lakes in WGS84 projection.
}
\details{
Note from Todd: "This polygon layer of GL shoreline was modified by
hand to include Saginaw, Tittabawasssee, Maumee, and Sandusky rivers.
Outlines of rivers are not precise but were wide enough to allow a
continuous connection between pixels for the entire undammed river stretch
when the 'rasterize' function is used to produce a raster layer of the GL
in QGIS."

Todd's original file name was 'coastline_poly_modified_rivers'.
}
\section{Filename}{
 shoreline.zip
}

\section{Used to make \link{great_lakes_polygon}.}{
NA
}

\examples{

# Read polygon from shapefile

poly_file <- system.file("extdata", "shoreline.zip", package = "glatos")

poly <- sf::st_read(paste0("/vsizip/", poly_file))

\dontrun{
plot(sf::st_geometry(poly))
}

}
\author{
Todd Hayden
}
