% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extdata.r
\name{shoreline}
\alias{shoreline}
\title{zipped polygon shapefile of Great Lakes}
\format{
shapefile
}
\source{
\url{http://glatos.glos.us/home}
}
\usage{
system.file("extdata", "shoreline.zip", package="glatos")
}
\description{
Polygon coastline of Great Lakes in WGS84
projection.  Includes outlines of Tittabawassee River (Lake
Huron), Maumee River (Lake Erie), and Sandusky River (Lake Erie)
}
\section{Filename}{
 shoreline.zip
}

\author{
Todd Hayden
}
