% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-write_vdat_csv.r
\name{[.vdat_list}
\alias{[.vdat_list}
\title{Subset method for vdat_list that retains attributes}
\usage{
\method{[}{vdat_list}(x, i, ...)
}
\arguments{
\item{x}{a vdat_list from which to extract element(s).}

\item{i}{indices specifying elements to extract or replace.}

\item{...}{unused.}
}
\description{
Subset method for vdat_list that retains attributes
}
