% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transmitAlongPath.r
\name{transmitAlongPath}
\alias{transmitAlongPath}
\title{Simulate telemetry transmitter signals along a path}
\usage{
transmitAlongPath(path = NA, vel = 0.5, delayRng = c(60, 180),
  burstDur = 5)
}
\arguments{
\item{path}{A two-column data frame with at least two rows and columns 
\code{x} and \code{y} with coordinates that define path.}

\item{vel}{A numeric scalar with movement velocity along track; assumed 
constant.}

\item{delayRng}{A 2-element numeric vector with minimum and maximum delay 
(time in seconds from end of one coded burst to beginning of next).}

\item{burstDur}{A numeric scalar with duration (in seconds) of each coded 
burst (i.e., pulse train).}
}
\value{
A two-column data frame containing:
  \item{x}{ x coordinates for start of each transmission }
  \item{y}{ y coordinates for start of each transmission }
  \item{et}{ elapsed time to start of each transmission }
}
\description{
Simulate tag signal transmission along a pre-defined path (x, y coords)
  based on constant movement velocity, transmitter delay range, and duration
  of signal.
}
\details{
Delays are drawn from uniform distribution defined by delay range.
First, elapsed time in seconds at each node in path is calculated based on 
path length and velocity. Next, delays are simulated and burst durations 
are added toeach delay to determine the time of each signal transmission. 
Location of each signal transmission along the path is linearly interpolated.
}
\note{
This function was written to be called before 
  \code{\link{detectTransmissions}}, which was designed to accept the result
  as input (\code{trnsLoc}).
}
\examples{
mypath <- data.frame(x=seq(0,1000,100),y=seq(0,1000,100))
mytrns <- transmitAlongPath(mypath,vel=0.5,delayRng=c(60,180),burstDur=5.0)
plot(mypath,type="b")
points(mytrns,pch=20,col="red")

}
\author{
C. Holbrook (cholbrook@usgs.gov)
}
