% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vectorHeading.r
\name{vectorHeading}
\alias{vectorHeading}
\title{Calculate direction (heading) of a vector (in degrees)}
\usage{
vectorHeading(x, y)
}
\arguments{
\item{x}{A numeric vector of x coordinates; minimum of 2.}

\item{y}{A numeric vector of y coordinates; minimum of 2.}
}
\value{
A numeric scalar with heading in degrees or a numeric vector of 
headings if \code{length(x) > 2}.
}
\description{
Calculate direction (heading) of a vector (in degrees)
}
\details{
Calculates direction (in degrees) for each of k-1 vectors, where 
k = length(x) - 1 = length(y). Lenghts of \code{x} and \code{y} must be 
equal.
}
\note{
This function is called from within \code{\link{crwInPolygon}}
}
\examples{
x=c(2,4)
y=c(2,4)
vectorHeading(x, y)

x2=c(2,4,2)
y2=c(2,4,2)
vectorHeading(x2, y2)

}
\author{
C. Holbrook (cholbrook@usgs.gov)
}
