% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filt-velTest.r
\name{velTest}
\alias{velTest}
\title{Check if measured velocity of a tagged animal is within plausible range}
\usage{
To use: For GLATOS data, velTest(data, "GLATOS") For OTN data,
 velTest(data, "OTN") For sample data, velTest(data, "sample")
}
\arguments{
\item{detections}{A data frame containing detection data with at least 5 
columns containing 'timestamp', 'transmitters', 'receivers', 'longitude', 
and 'latitude'. Column names are specified by \code{detColNames} or 
indirectly by \code{type} (for GLATOS and OTN data).}

\item{type}{A character string with the type of detection data
passed in, if applicable. Acceptable values are "GLATOS" (default) or "OTN".
 Not used (ignored) when \code{detColNames} is used.}

\item{detColNames}{An optional list that contains the user-defined column 
names of reuired columns in \code{detections}. See Details.}

\item{maxVelocity}{The maximum expected velocity ('ground speed') of a tagged 
animal, in meters per second.}
}
\value{
A data frame containing the data with the four columns appended to it:
 \item{min_dist}{Minimum of the distance between the current instance and
 instance before, and the distance between the current instance and the
 instance after.} \item{min_time}{Minimum of the time between the current
 instance and instance before, and the time between the current instance and
 the instance after.} \item{min_vel}{The min_dist value divided by the
 min_time value.} \item{velValid}{A value to check if the min_vel value is
 greater than or equal to (0), or less than (1) the \code{maxVelocity}.}
}
\description{
Perform velocity test (Steckenreuter et al., 2017), which checks if 
measured velocity exceeds a threshold (maximum velocity).
}
\details{
detColNames A list with the names of the required
 columns in \code{detections}: \itemize{ \item \code{timestampCol} is a
 character string with the name of the column containing datetime stamps for
 the detections (MUST be of class 'POSIXct') ('detection_timestamp_utc' for
 GLATOS data, 'datecollected' for OTN data, or 'time' for sample data). \item
 \code{transmittersCol} is a character string with the name of the column 
 containing the ids of the transmitters ('transmission_id' for GLATOS data,
 'tagname' for OTN data, or 'transmitter' for sample data). \item
 \code{receiversCol} is a character string with the name of the column 
 containing the ids of the receivers ('receiver_sn' for GLATOS data,
 'receiver_group' for OTN data, or 'receiver' for sample data). \item
 \code{longitudeCol} is a character string with the name of the column 
 containing the longitude coordinate for the detections ('deploy_long' for
 GLATOS data, 'longitude' for OTN data, or 'longitude' for sample data). 
 \item \code{latitudeCol} is a character string with the name of the column 
 containing the latitude coordinate for the detections ('deploy_lat' for
 GLATOS data, 'latitude' for OTN data, or 'latitude' for sample data). }

Each value in the min_dist column indicates the minimum of the
 distance between the current instance and instance before, and the distance
 between the current instance and the instance after.

Each value in the min_time column indicates the minimum of the time
 between the current instance and instance before, and the time between the
 current instance and the instance after.

Each value in the min_vel column indicates the value in the min_dist
 column divided by the value in the min_time column.

Each value in the velValid column indicates whether the value in the
 min_vel column is greater than the threshold 'maxVelocity'.
}
\references{
(in APA) Steckenreuter, A., Hoenner, X., Huveneers, C.,
 Simpfendorfer, C., Buscot, M.J., Tattersall, K., ... Harcourt, R. (2017).
 Optimising the design of large-scale acoustic telemetry curtains. Marine and
 Freshwater Research. 68:1403-1413. doi: 10.1071/MF16126
}
\author{
A. Dini
}
