% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/walleye_controlTable.r
\docType{data}
\name{walleye_controlTable}
\alias{walleye_controlTable}
\title{An example control data frame that determine order of location groups on the 
y-axis of an \link{abacusPlot} or \link{eventPlot}}
\description{
An example control data frame that determine order of location groups on the 
y-axis of an \link{abacusPlot} or \link{eventPlot}. This file is an example 
provided by Todd Hayden for use with example data \link{walleye_detections} 
with the functions \link{abacusPlot} and \link{eventPlot}.
}
\seealso{
\link{abacusPlot}, \link{eventPlot}, \link{walleye_detections}
}
\author{
Todd Hayden
}
