% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{walleye_detections}
\alias{walleye_detections}
\title{A data frame containing fish detection data}
\format{An object of class \code{data.frame} with 150926 rows and 30 columns.}
\usage{
walleye_detections
}
\description{
A data frame containing fish detection data from the GLATOS Project 
HECSL. This is a sample of detection data in the standard 
GLATOS export format.
}
\author{
Todd Hayden
}
\keyword{datasets}
