% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r, R/walleye_plotControl.r
\docType{data}
\name{walleye_plotControl}
\alias{walleye_plotControl}
\alias{walleye_plotControl}
\title{An example plot control data frame that determines the color and marker of 
points png and mp4 files created by \link{animatePath}}
\format{An object of class \code{data.frame} with 20 rows and 4 columns.}
\usage{
walleye_plotControl
}
\description{
An example plot control data frame that determines the color and marker of 
points png and mp4 files created by \link{animatePath}. Provided by Todd 
Hayden for use with example file \link{walleye_detections}
with the functions \link{interpolatePath} and \link{animatePath}.

An example plot control data frame that determines the color and marker of 
points png and mp4 files created by \link{animatePath}. Provided by Todd 
Hayden for use with example file \link{walleye_detections}
with the functions \link{interpolatePath} and \link{animatePath}.
}
\seealso{
\link{interpolatePath}, \link{animatePath}, 
\link{walleye_detections}

\link{interpolatePath}, \link{animatePath}, 
\link{walleye_detections}
}
\author{
Todd Hayden

Todd Hayden
}
\keyword{datasets}
